Consolidate results from partial files

Consolidates results from a set of partial files (each generated by an individual call to calc_nreps()) into a single output structure, similar (but not identical) to the output of run_experiment(). This is useful e.g., to consolidate the results from interrupted experiments.

consolidate_partial_results(Configuration, folder = "./nreps_files")

Arguments

Configuration
a named list containing all parameters required in a call to run_experiment() except instances and algorithms. See the parameter list and default values in run_experiment(). Notice that this is always returned as part of the output structure of run_experiment(), so it generally easier to just retrieve it from previously saved results.

folder
folder where the partial files are located.

Value

a list object containing the following fields:

data.raw - data frame containing all observations generated

data.summary - data frame summarizing the experiment.

N - number of instances sampled

total.runs - total number of algorithm runs performed

instances.sampled - names of the instances sampled