print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ...
#> Sampling algorithms on instance Inst.5 : ...
#> Sampling algorithms on instance Inst.38 : ...
#> Sampling algorithms on instance Inst.16 : ...
#> Sampling algorithms on instance Inst.4 : ...
#> Sampling algorithms on instance Inst.86 : ...
#> Sampling algorithms on instance Inst.90 : ...
#> Sampling algorithms on instance Inst.70 : ...
#> Sampling algorithms on instance Inst.79 : ...
#> Sampling algorithms on instance Inst.78 : ...
#> Sampling algorithms on instance Inst.14 : ...
#> Sampling algorithms on instance Inst.56 : ...
#> Sampling algorithms on instance Inst.62 : ...
#> Sampling algorithms on instance Inst.93 : ...

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.41063 0.05704
#>   Inst.28 algo1 algo3 20 33 -0.79398 0.07338
#>   Inst.28 algo1 algo4 20 40  0.03933 0.07509
#>   Inst.28 algo2 algo3 20 33 -1.20462 0.08903
#>   Inst.28 algo2 algo4 20 40 -0.37130 0.08746
#>   Inst.28 algo3 algo4 33 40  0.83331 0.09926
#>   Inst.80 algo1 algo2 20 20  0.40941 0.05698
#>   Inst.80 algo1 algo3 20 32 -0.78668 0.07163
#>   Inst.80 algo1 algo4 20 42  0.01656 0.07642
#>   Inst.80 algo2 algo3 20 32 -1.19609 0.08750
#>   Inst.80 algo2 algo4 20 42 -0.39285 0.08863
#>   Inst.80 algo3 algo4 32 42  0.80324 0.09884
#>   Inst.22 algo1 algo2 20 20  0.41018 0.05700
#>   Inst.22 algo1 algo3 20 33 -0.79736 0.07334
#>   Inst.22 algo1 algo4 20 40  0.03711 0.07540
#>   Inst.22 algo2 algo3 20 33 -1.20753 0.08899
#>   Inst.22 algo2 algo4 20 40 -0.37307 0.08771
#>   Inst.22 algo3 algo4 33 40  0.83447 0.09946
#>    Inst.9 algo1 algo2 20 20  0.31742 0.06802
#>    Inst.9 algo1 algo3 20 37 -0.89893 0.07202
#>    Inst.9 algo1 algo4 20 49  0.07265 0.07703
#>    Inst.9 algo2 algo3 20 37 -1.21634 0.09501
#>    Inst.9 algo2 algo4 20 49 -0.24476 0.09616
#>    Inst.9 algo3 algo4 37 49  0.97158 0.09979
#>    Inst.5 algo1 algo2 20 20  0.31948 0.06808
#>    Inst.5 algo1 algo3 20 37 -0.89603 0.07219
#>    Inst.5 algo1 algo4 20 49  0.07385 0.07723
#>    Inst.5 algo2 algo3 20 37 -1.21552 0.09506
#>    Inst.5 algo2 algo4 20 49 -0.24564 0.09622
#>    Inst.5 algo3 algo4 37 49  0.96988 0.09994
#>   Inst.38 algo1 algo2 20 20  0.31906 0.06807
#>   Inst.38 algo1 algo3 20 37 -0.89663 0.07215
#>   Inst.38 algo1 algo4 20 49  0.07360 0.07719
#>   Inst.38 algo2 algo3 20 37 -1.21569 0.09505
#>   Inst.38 algo2 algo4 20 49 -0.24546 0.09621
#>   Inst.38 algo3 algo4 37 49  0.97023 0.09991
#>   Inst.16 algo1 algo2 20 20  0.31704 0.06802
#>   Inst.16 algo1 algo3 20 37 -0.89946 0.07200
#>   Inst.16 algo1 algo4 20 49  0.07244 0.07699
#>   Inst.16 algo2 algo3 20 37 -1.21650 0.09501
#>   Inst.16 algo2 algo4 20 49 -0.24460 0.09615
#>   Inst.16 algo3 algo4 37 49  0.97190 0.09976
#>    Inst.4 algo1 algo2 20 20  0.31641 0.06801
#>    Inst.4 algo1 algo3 20 37 -0.90034 0.07196
#>    Inst.4 algo1 algo4 20 49  0.07208 0.07694
#>    Inst.4 algo2 algo3 20 37 -1.21675 0.09500
#>    Inst.4 algo2 algo4 20 49 -0.24434 0.09613
#>    Inst.4 algo3 algo4 37 49  0.97241 0.09972
#>   Inst.86 algo1 algo2 20 20  0.32241 0.06846
#>   Inst.86 algo1 algo3 20 37 -0.89840 0.07263
#>   Inst.86 algo1 algo4 20 51  0.09424 0.07634
#>   Inst.86 algo2 algo3 20 37 -1.22081 0.09551
#>   Inst.86 algo2 algo4 20 51 -0.22817 0.09554
#>   Inst.86 algo3 algo4 37 51  0.99264 0.09956
#>   Inst.90 algo1 algo2 20 20  0.31997 0.06810
#>   Inst.90 algo1 algo3 20 37 -0.89536 0.07223
#>   Inst.90 algo1 algo4 20 49  0.07412 0.07729
#>   Inst.90 algo2 algo3 20 37 -1.21532 0.09507
#>   Inst.90 algo2 algo4 20 49 -0.24584 0.09624
#>   Inst.90 algo3 algo4 37 49  0.96948 0.09998
#>   Inst.70 algo1 algo2 20 20  0.31404 0.07178
#>   Inst.70 algo1 algo3 20 35 -0.93044 0.07414
#>   Inst.70 algo1 algo4 20 47  0.11188 0.08010
#>   Inst.70 algo2 algo3 20 35 -1.24448 0.09506
#>   Inst.70 algo2 algo4 20 47 -0.20215 0.09685
#>   Inst.70 algo3 algo4 35 47  1.04233 0.09985
#>   Inst.79 algo1 algo2 20 20  0.30830 0.07209
#>   Inst.79 algo1 algo3 20 35 -0.87540 0.07631
#>   Inst.79 algo1 algo4 20 47  0.04632 0.07885
#>   Inst.79 algo2 algo3 20 35 -1.18371 0.09687
#>   Inst.79 algo2 algo4 20 47 -0.26198 0.09628
#>   Inst.79 algo3 algo4 35 47  0.92172 0.09997
#>   Inst.78 algo1 algo2 20 20  0.31150 0.07181
#>   Inst.78 algo1 algo3 20 34 -0.93982 0.07519
#>   Inst.78 algo1 algo4 20 48  0.11443 0.07870
#>   Inst.78 algo2 algo3 20 34 -1.25132 0.09594
#>   Inst.78 algo2 algo4 20 48 -0.19708 0.09576
#>   Inst.78 algo3 algo4 34 48  1.05424 0.09958
#>   Inst.14 algo1 algo2 20 20  0.31408 0.07178
#>   Inst.14 algo1 algo3 20 35 -0.93038 0.07413
#>   Inst.14 algo1 algo4 20 47  0.11191 0.08011
#>   Inst.14 algo2 algo3 20 35 -1.24447 0.09505
#>   Inst.14 algo2 algo4 20 47 -0.20217 0.09685
#>   Inst.14 algo3 algo4 35 47  1.04229 0.09985
#>   Inst.56 algo1 algo2 20 20  0.31296 0.07177
#>   Inst.56 algo1 algo3 20 34 -0.92648 0.07526
#>   Inst.56 algo1 algo4 20 48  0.10425 0.07905
#>   Inst.56 algo2 algo3 20 34 -1.23944 0.09593
#>   Inst.56 algo2 algo4 20 48 -0.20871 0.09602
#>   Inst.56 algo3 algo4 34 48  1.03073 0.09981
#>   Inst.62 algo1 algo2 20 20  0.31133 0.07182
#>   Inst.62 algo1 algo3 20 34 -0.93999 0.07521
#>   Inst.62 algo1 algo4 20 48  0.11428 0.07869
#>   Inst.62 algo2 algo3 20 34 -1.25132 0.09596
#>   Inst.62 algo2 algo4 20 48 -0.19705 0.09575
#>   Inst.62 algo3 algo4 34 48  1.05427 0.09958
#>   Inst.93 algo1 algo2 20 20  0.31494 0.07168
#>   Inst.93 algo1 algo3 20 34 -0.93531 0.07525
#>   Inst.93 algo1 algo4 20 48  0.11627 0.07872
#>   Inst.93 algo2 algo3 20 34 -1.25025 0.09587
#>   Inst.93 algo2 algo4 20 48 -0.19867 0.09563
#>   Inst.93 algo3 algo4 34 48  1.05158 0.09962
#> 
#> #====================================