Run a full experiment for comparing multiple algorithms using multiple instances

Design and run a full experiment - calculate the required number of instances, run the algorithms on each problem instance using the iterative approach based on optimal sample size ratios, and return the results of the experiment. This routine builds upon calc_instances() and calc_nreps(), so refer to the documentation of these two functions for details.

run_experiment(instances, algorithms, d, se.max, power = 0.8,
  sig.level = 0.05, power.target = "mean", dif = "simple",
  comparisons = "all.vs.all", alternative = "two.sided",
  test = "t.test", method = "param", nstart = 20, nmax = 100 *
  length(algorithms), force.balanced = FALSE, ncpus = 2,
  boot.R = 499, seed = NULL, save.partial.results = NA,
  load.partial.results = NA, save.final.result = NA)

Arguments

instances
list object containing the definitions of the available instances. This list may (or may not) be exhausted in the experiment. To estimate the number of required instances, see calc_instances(). For more details, see Section Instance List.

algorithms
a list object containing the definitions of all algorithms. See Section Algorithms for details.

d
minimally relevant effect size (MRES), expressed as a standardized effect size, i.e., "deviation from H0" / "standard deviation". See calc_instances() for details.

se.max
desired upper limit for the standard error of the estimated difference between pairs of algorithms. See Section Pairwise Differences for details.

power
(desired) test power. See calc_instances() for details. Any value equal to or greater than one will force the method to use all available instances in Instance.list.

sig.level
family-wise significance level (alpha) for the experiment. See calc_instances() for details.

power.target
which comparison should have the desired power? Accepts "mean", "median", or "worst.case" (this last one is equivalent to the Bonferroni correction).

dif
type of difference to be used. Accepts "perc" (for percent differences) or "simple" (for simple differences)

comparisons
type of comparisons being performed. Accepts "all.vs.first" (in which cases the first object in algorithms is considered to be the reference algorithm) or "all.vs.all" (if there is no reference and all pairwise comparisons are desired).

alternative
type of alternative hypothesis ("two.sided" or "less" or "greater"). See calc_instances() for details.

test
type of test to be used ("t.test", "wilcoxon" or "binomial")

method
method to use for estimating the standard errors. Accepts "param" (for parametric) or "boot" (for bootstrap)

nstart
initial number of algorithm runs for each algorithm. See Section Initial Number of Observations for details.

nmax
maximum number of runs to execute on each instance (see calc_nreps()). Loaded results (see load.partial.results below) do not count towards this maximum.

force.balanced
logical flag to force the use of balanced sampling for the algorithms on each instance

ncpus
number of cores to use

boot.R
number of bootstrap resamples to use (if method == "boot")

seed
seed for the random number generator

save.partial.results
should partial results be saved to files? Can be either NA (do not save) or a character string pointing to a folder. File names are generated based on the instance aliases. Existing files with matching names will be overwritten. run_experiment() uses .RDS files for saving and loading.

load.partial.results
should partial results be loaded from files? Can be either NA (do not save) or a character string pointing to a folder containing the file(s) to be loaded. run_experiment() will use .RDS file(s) with a name(s) matching instance aliases. run_experiment() uses .RDS files for saving and loading.

save.final.result
should the final results be saved to file? Can be either NA (do not save) or a character string pointing to a folder where the results will be saved on a .RDS file starting with CAISEr_results_ and ending with 12-digit datetime tag in the format YYYYMMDDhhmmss.

Value

a list object containing the following fields:

Configuration - the full input configuration (for reproducibility)

data.raw - data frame containing all observations generated

data.summary - data frame summarizing the experiment.

N - number of instances sampled

N.star - number of instances required

total.runs - total number of algorithm runs performed

instances.sampled - names of the instances sampled

Underpowered - flag: TRUE if N < N.star

Instance List

Parameter instances must contain a list of instance objects, where each field is itself a list, as defined in the documentation of function calc_nreps(). In short, each element of instances is an instance, i.e., a named list containing all relevant parameters that define the problem instance. This list must contain at least the field instance$FUN, with the name of the problem instance function, that is, a routine that calculates y = f(x). If the instance requires additional parameters, these must also be provided as named fields. An additional field, "instance$alias", can be used to provide the instance with a unique identifier (e.g., when using an instance generator).

Algorithm List

Object algorithms is a list in which each component is a named list containing all relevant parameters that define an algorithm to be applied for solving the problem instance. In what follows algorithms[[k]] refers to any algorithm specified in the algorithms list.

algorithms[[k]] must contain an algorithms[[k]]$FUN field, which is a character object with the name of the function that calls the algorithm; as well as any other elements/parameters that algorithms[[k]]$FUN requires (e.g., stop criteria, operator names and parameters, etc.).

The function defined by the routine algorithms[[k]]$FUN must have the following structure: supposing that the list in algorithms[[k]] has fields algorithm[[k]]$FUN = "myalgo", algorithms[[k]]$par1 = "a" and algorithms[[k]]$par2 = 5, then:

         myalgo <- function(par1, par2, instance, ...){
               #
               # <do stuff>
               #
               return(results)
         }
   

That is, it must be able to run if called as:

         # remove '$FUN' and '$alias' field from list of arguments
         # and include the problem definition as field 'instance'
         myargs          <- algorithm[names(algorithm) != "FUN"]
         myargs          <- myargs[names(myargs) != "alias"]
         myargs$instance <- instance
    # call function
         do.call(algorithm$FUN,
                 args = myargs)
   

The algorithm$FUN routine must return a list containing (at least) the performance value of the final solution obtained, in a field named value (e.g., result$value) after a given run. In general it is easier to write a small wrapper function around existing implementations.

Initial Number of Observations

In the general case the initial number of observations / algorithm / instance (nstart) should be relatively high. For the parametric case we recommend 10~15 if outliers are not expected, and 30~40 (at least) if that assumption cannot be made. For the bootstrap approach we recommend using at least 15 or 20. However, if some distributional assumptions can be made - particularly low skewness of the population of algorithm results on the test instances), then nstart can in principle be as small as 5 (if the output of the algorithm were known to be normal, it could be 1).

In general, higher sample sizes are the price to pay for abandoning distributional assumptions. Use lower values of nstart with caution.

Pairwise Differences

Parameter dif informs the type of difference in performance to be used for the estimation (\(\mu_a\) and \(\mu_b\) represent the mean performance of any two algorithms on the test instance, and \(mu\) represents the grand mean of all algorithms given in algorithms):

If dif == "perc" and comparisons == "all.vs.first", the estimated quantity is: \(\phi_{1b} = (\mu_1 - \mu_b) / \mu_1 = 1 - (\mu_b / \mu_1)\).

If dif == "perc" and comparisons == "all.vs.all", the estimated quantity is: \(\phi_{ab} = (\mu_a - \mu_b) / \mu\).

If dif == "simple" it estimates \(\mu_a - \mu_b\).

Sample Sizes for Nonparametric Methods

If the parameter `` is set to either Wilcoxon or `Binomial`, this routine approximates the number of instances using the ARE of these tests in relation to the paired t.test:

n.wilcox = n.ttest / 0.86 = 1.163 * n.ttest

n.binom = n.ttest / 0.637 = 1.570 * n.ttest

References

F. Campelo, F. Takahashi: Sample size estimation for power and accuracy in the experimental comparison of algorithms. Journal of Heuristics 25(2):305-338, 2019.

P. Mathews. Sample size calculations: Practical methods for engineers and scientists. Mathews Malnar and Bailey, 2010.

A.C. Davison, D.V. Hinkley: Bootstrap methods and their application. Cambridge University Press (1997)

E.C. Fieller: Some problems in interval estimation. Journal of the Royal Statistical Society. Series B (Methodological) 16(2), 175–185 (1954)

V. Franz: Ratios: A short guide to confidence limits and proper use (2007). https://arxiv.org/pdf/0710.2024v1.pdf

D.C. Montgomery, C.G. Runger: Applied Statistics and Probability for Engineers, 6th ed. Wiley (2013)

D.J. Sheskin: Handbook of Parametric and Nonparametric Statistical Procedures, 4th ed., Chapman & Hall/CRC, 1996.

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = .5, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             ncpus = 1, seed = 1234)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 58
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..
#> Sampling algorithms on instance Inst.84 : ..
#> Sampling algorithms on instance Inst.21 : ..
#> Sampling algorithms on instance Inst.40 : ..
#> Sampling algorithms on instance Inst.92 : ..
#> Sampling algorithms on instance Inst.67 : ..
#> Sampling algorithms on instance Inst.96 : ..
#> Sampling algorithms on instance Inst.66 : ..
#> Sampling algorithms on instance Inst.47 : ..
#> Sampling algorithms on instance Inst.81 : ..
#> Sampling algorithms on instance Inst.48 : ..
#> Sampling algorithms on instance Inst.3 : ..
#> Sampling algorithms on instance Inst.41 : ..
#> Sampling algorithms on instance Inst.32 : ..
#> Sampling algorithms on instance Inst.42 : ..
#> Sampling algorithms on instance Inst.43 : ..
#> Sampling algorithms on instance Inst.2 : ..
#> Sampling algorithms on instance Inst.54 : ..
#> Sampling algorithms on instance Inst.49 : ..
#> Sampling algorithms on instance Inst.99 : ..
#> Sampling algorithms on instance Inst.51 : ..
#> Sampling algorithms on instance Inst.6 : ..
#> Sampling algorithms on instance Inst.77 : ..
#> Sampling algorithms on instance Inst.29 : ..
#> Sampling algorithms on instance Inst.71 : ..
#> Sampling algorithms on instance Inst.85 : ..
#> Sampling algorithms on instance Inst.57 : ..
#> Sampling algorithms on instance Inst.8 : ..
#> Sampling algorithms on instance Inst.26 : ..
#> Sampling algorithms on instance Inst.17 : ..
#> Sampling algorithms on instance Inst.58 : ..
#> Sampling algorithms on instance Inst.91 : ..
#> Sampling algorithms on instance Inst.60 : ..
#> Sampling algorithms on instance Inst.76 : ..
#> Sampling algorithms on instance Inst.98 : ..
#> Sampling algorithms on instance Inst.35 : ..
#> Sampling algorithms on instance Inst.72 : ..
#> Sampling algorithms on instance Inst.94 : ..
#> Sampling algorithms on instance Inst.73 : ..
#> Sampling algorithms on instance Inst.50 : ..
#> Sampling algorithms on instance Inst.10 : ..
#> Sampling algorithms on instance Inst.45 : ..

# Take a look at the results
summary(my.results)

#> #====================================
#>  CAISEr object:
#>  Number of instances sampled: 58
#>  Number of instances required: 58
#>  Adequate power: TRUE
#>  Total runs of algo1 : 1160
#>  Total runs of algo2 : 1160
#>  Total runs of algo3 : 1640
#>  Total runs of algo4 : 2147
#> #====================================
#>  Pairwise comparisons of interest:
#>  Test: t.test
#>  H1: two.sided
#>  Comparisons: all.vs.all
#>  Alpha (FWER): 0.05
#>  Power target: mean
#>  Desired power: 0.9
#> #====================================
#> Tests using Holm's step-down procedure:
#> 
#>  Test 1 : algo1 - algo3
#>  H0: mean = 0
#>  alpha		= 0.008333 
#>  p-value	= 8.421e-61
#>  Est. mean	= -0.8854
#>  CI{1-alpha}	= [ -0.915 -0.8558 ]
#>  d		= -10.74481
#> 
#>  Test 2 : algo2 - algo3
#>  H0: mean = 0
#>  alpha		= 0.01 
#>  p-value	= 2.074e-58
#>  Est. mean	= -1.118
#>  CI{1-alpha}	= [ -1.158 -1.078 ]
#>  d		= -9.746685
#> 
#>  Test 3 : algo3 - algo4
#>  H0: mean = 0
#>  alpha		= 0.0125 
#>  p-value	= 5.234e-45
#>  Est. mean	= 0.851
#>  CI{1-alpha}	= [ 0.7996 0.9023 ]
#>  d		= 5.615371
#> 
#>  Test 4 : algo1 - algo2
#>  H0: mean = 0
#>  alpha		= 0.01667 
#>  p-value	= 2.112e-42
#>  Est. mean	= 0.2325
#>  CI{1-alpha}	= [ 0.2176 0.2475 ]
#>  d		= 5.036149
#> 
#>  Test 5 : algo2 - algo4
#>  H0: mean = 0
#>  alpha		= 0.025 
#>  p-value	= 1.137e-34
#>  Est. mean	= -0.267
#>  CI{1-alpha}	= [ -0.2892 -0.2447 ]
#>  d		= -3.623905
#> 
#>  Test 6 : algo1 - algo4
#>  H0: mean = 0
#>  alpha		= 0.05 
#>  p-value	= 0.0007322
#>  Est. mean	= -0.03443
#>  CI{1-alpha}	= [ -0.05374 -0.01512 ]
#>  d		= -0.4688137
#> #====================================

print(my.results)

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2       Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.300579 0.05017
#>   Inst.28 algo1 algo3 20 24 -0.907237 0.07262
#>   Inst.28 algo1 algo4 20 39 -0.005947 0.08131
#>   Inst.28 algo2 algo3 20 24 -1.207816 0.07937
#>   Inst.28 algo2 algo4 20 39 -0.306526 0.08507
#>   Inst.28 algo3 algo4 24 39  0.901289 0.09991
#>   Inst.80 algo1 algo2 20 20  0.300409 0.05017
#>   Inst.80 algo1 algo3 20 24 -0.907297 0.07258
#>   Inst.80 algo1 algo4 20 39 -0.006385 0.08126
#>   Inst.80 algo2 algo3 20 24 -1.207706 0.07939
#>   Inst.80 algo2 algo4 20 39 -0.306795 0.08508
#>   Inst.80 algo3 algo4 24 39  0.900912 0.09989
#>   Inst.22 algo1 algo2 20 20  0.298741 0.05036
#>   Inst.22 algo1 algo3 20 25 -0.888685 0.07191
#>   Inst.22 algo1 algo4 20 38 -0.028316 0.08136
#>   Inst.22 algo2 algo3 20 25 -1.187425 0.07952
#>   Inst.22 algo2 algo4 20 38 -0.327056 0.08596
#>   Inst.22 algo3 algo4 25 38  0.860369 0.09987
#>    Inst.9 algo1 algo2 20 20  0.300346 0.05017
#>    Inst.9 algo1 algo3 20 24 -0.907320 0.07256
#>    Inst.9 algo1 algo4 20 39 -0.006549 0.08124
#>    Inst.9 algo2 algo3 20 24 -1.207666 0.07940
#>    Inst.9 algo2 algo4 20 39 -0.306895 0.08509
#>    Inst.9 algo3 algo4 24 39  0.900771 0.09988
#>    Inst.5 algo1 algo2 20 20  0.300547 0.05017
#>    Inst.5 algo1 algo3 20 24 -0.907248 0.07261
#>    Inst.5 algo1 algo4 20 39 -0.006031 0.08130
#>    Inst.5 algo2 algo3 20 24 -1.207795 0.07937
#>    Inst.5 algo2 algo4 20 39 -0.306578 0.08507
#>    Inst.5 algo3 algo4 24 39  0.901217 0.09991
#>   Inst.38 algo1 algo2 20 20  0.300277 0.05017
#>   Inst.38 algo1 algo3 20 24 -0.907344 0.07254
#>   Inst.38 algo1 algo4 20 39 -0.006729 0.08122
#>   Inst.38 algo2 algo3 20 24 -1.207621 0.07941
#>   Inst.38 algo2 algo4 20 39 -0.307005 0.08509
#>   Inst.38 algo3 algo4 24 39  0.900616 0.09987
#>   Inst.16 algo1 algo2 20 20  0.300437 0.05017
#>   Inst.16 algo1 algo3 20 24 -0.907287 0.07258
#>   Inst.16 algo1 algo4 20 39 -0.006313 0.08127
#>   Inst.16 algo2 algo3 20 24 -1.207724 0.07939
#>   Inst.16 algo2 algo4 20 39 -0.306750 0.08508
#>   Inst.16 algo3 algo4 24 39  0.900974 0.09989
#>    Inst.4 algo1 algo2 20 20  0.300387 0.05017
#>    Inst.4 algo1 algo3 20 24 -0.907305 0.07257
#>    Inst.4 algo1 algo4 20 39 -0.006443 0.08125
#>    Inst.4 algo2 algo3 20 24 -1.207692 0.07939
#>    Inst.4 algo2 algo4 20 39 -0.306830 0.08508
#>    Inst.4 algo3 algo4 24 39  0.900862 0.09988
#>   Inst.86 algo1 algo2 20 20  0.299891 0.05018
#>   Inst.86 algo1 algo3 20 24 -0.907482 0.07244
#>   Inst.86 algo1 algo4 20 39 -0.007724 0.08111
#>   Inst.86 algo2 algo3 20 24 -1.207373 0.07947
#>   Inst.86 algo2 algo4 20 39 -0.307615 0.08513
#>   Inst.86 algo3 algo4 24 39  0.899758 0.09981
#>   Inst.90 algo1 algo2 20 20  0.300058 0.05017
#>   Inst.90 algo1 algo3 20 24 -0.907422 0.07248
#>   Inst.90 algo1 algo4 20 39 -0.007292 0.08115
#>   Inst.90 algo2 algo3 20 24 -1.207480 0.07944
#>   Inst.90 algo2 algo4 20 39 -0.307351 0.08511
#>   Inst.90 algo3 algo4 24 39  0.900130 0.09984
#>   Inst.70 algo1 algo2 20 20  0.298077 0.05050
#>   Inst.70 algo1 algo3 20 25 -0.888978 0.07177
#>   Inst.70 algo1 algo4 20 38 -0.030018 0.08120
#>   Inst.70 algo2 algo3 20 25 -1.187054 0.07968
#>   Inst.70 algo2 algo4 20 38 -0.328094 0.08608
#>   Inst.70 algo3 algo4 25 38  0.858960 0.09979
#>   Inst.79 algo1 algo2 20 20  0.187329 0.05164
#>   Inst.79 algo1 algo3 20 29 -0.769263 0.07206
#>   Inst.79 algo1 algo4 20 34 -0.129081 0.07745
#>   Inst.79 algo2 algo3 20 29 -0.956592 0.08364
#>   Inst.79 algo2 algo4 20 34 -0.316410 0.08728
#>   Inst.79 algo3 algo4 29 34  0.640182 0.09977
#>   Inst.78 algo1 algo2 20 20  0.189685 0.05164
#>   Inst.78 algo1 algo3 20 29 -0.768108 0.07200
#>   Inst.78 algo1 algo4 20 34 -0.128654 0.07751
#>   Inst.78 algo2 algo3 20 29 -0.957793 0.08367
#>   Inst.78 algo2 algo4 20 34 -0.318339 0.08739
#>   Inst.78 algo3 algo4 29 34  0.639453 0.09983
#>   Inst.14 algo1 algo2 20 20  0.188910 0.05161
#>   Inst.14 algo1 algo3 20 29 -0.768488 0.07201
#>   Inst.14 algo1 algo4 20 34 -0.128795 0.07748
#>   Inst.14 algo2 algo3 20 29 -0.957398 0.08364
#>   Inst.14 algo2 algo4 20 34 -0.317705 0.08734
#>   Inst.14 algo3 algo4 29 34  0.639693 0.09980
#>   Inst.56 algo1 algo2 20 20  0.189003 0.05162
#>   Inst.56 algo1 algo3 20 29 -0.768442 0.07201
#>   Inst.56 algo1 algo4 20 34 -0.128778 0.07748
#>   Inst.56 algo2 algo3 20 29 -0.957445 0.08365
#>   Inst.56 algo2 algo4 20 34 -0.317781 0.08735
#>   Inst.56 algo3 algo4 29 34  0.639664 0.09981
#>   Inst.62 algo1 algo2 20 20  0.185362 0.05183
#>   Inst.62 algo1 algo3 20 29 -0.770227 0.07219
#>   Inst.62 algo1 algo4 20 34 -0.129437 0.07746
#>   Inst.62 algo2 algo3 20 29 -0.955589 0.08373
#>   Inst.62 algo2 algo4 20 34 -0.314799 0.08728
#>   Inst.62 algo3 algo4 29 34  0.640790 0.09976
#>   Inst.93 algo1 algo2 20 20  0.189023 0.05162
#>   Inst.93 algo1 algo3 20 29 -0.768432 0.07201
#>   Inst.93 algo1 algo4 20 34 -0.128774 0.07748
#>   Inst.93 algo2 algo3 20 29 -0.957456 0.08365
#>   Inst.93 algo2 algo4 20 34 -0.317798 0.08735
#>   Inst.93 algo3 algo4 29 34  0.639658 0.09981
#>   Inst.84 algo1 algo2 20 20  0.189647 0.05164
#>   Inst.84 algo1 algo3 20 29 -0.768126 0.07200
#>   Inst.84 algo1 algo4 20 34 -0.128661 0.07751
#>   Inst.84 algo2 algo3 20 29 -0.957774 0.08367
#>   Inst.84 algo2 algo4 20 34 -0.318309 0.08739
#>   Inst.84 algo3 algo4 29 34  0.639465 0.09983
#>   Inst.21 algo1 algo2 20 20  0.181999 0.05255
#>   Inst.21 algo1 algo3 20 29 -0.771876 0.07260
#>   Inst.21 algo1 algo4 20 34 -0.130045 0.07763
#>   Inst.21 algo2 algo3 20 29 -0.953875 0.08410
#>   Inst.21 algo2 algo4 20 34 -0.312045 0.08745
#>   Inst.21 algo3 algo4 29 34  0.641830 0.09982
#>   Inst.40 algo1 algo2 20 20  0.189519 0.05163
#>   Inst.40 algo1 algo3 20 29 -0.768189 0.07200
#>   Inst.40 algo1 algo4 20 34 -0.128684 0.07750
#>   Inst.40 algo2 algo3 20 29 -0.957708 0.08366
#>   Inst.40 algo2 algo4 20 34 -0.318204 0.08738
#>   Inst.40 algo3 algo4 29 34  0.639505 0.09982
#>   Inst.92 algo1 algo2 20 20  0.187568 0.05163
#>   Inst.92 algo1 algo3 20 29 -0.769146 0.07205
#>   Inst.92 algo1 algo4 20 34 -0.129038 0.07745
#>   Inst.92 algo2 algo3 20 29 -0.956714 0.08364
#>   Inst.92 algo2 algo4 20 34 -0.316606 0.08729
#>   Inst.92 algo3 algo4 29 34  0.640108 0.09977
#>   Inst.67 algo1 algo2 20 20  0.188691 0.05161
#>   Inst.67 algo1 algo3 20 29 -0.768595 0.07201
#>   Inst.67 algo1 algo4 20 34 -0.128834 0.07747
#>   Inst.67 algo2 algo3 20 29 -0.957286 0.08364
#>   Inst.67 algo2 algo4 20 34 -0.317526 0.08733
#>   Inst.67 algo3 algo4 29 34  0.639761 0.09980
#>   Inst.96 algo1 algo2 20 20  0.189814 0.05165
#>   Inst.96 algo1 algo3 20 29 -0.768045 0.07201
#>   Inst.96 algo1 algo4 20 34 -0.128631 0.07752
#>   Inst.96 algo2 algo3 20 29 -0.957859 0.08367
#>   Inst.96 algo2 algo4 20 34 -0.318445 0.08740
#>   Inst.96 algo3 algo4 29 34  0.639413 0.09983
#>   Inst.66 algo1 algo2 20 20  0.189317 0.05163
#>   Inst.66 algo1 algo3 20 29 -0.768288 0.07201
#>   Inst.66 algo1 algo4 20 34 -0.128721 0.07750
#>   Inst.66 algo2 algo3 20 29 -0.957605 0.08366
#>   Inst.66 algo2 algo4 20 34 -0.318038 0.08737
#>   Inst.66 algo3 algo4 29 34  0.639567 0.09982
#>   Inst.47 algo1 algo2 20 20  0.189225 0.05162
#>   Inst.47 algo1 algo3 20 29 -0.768333 0.07201
#>   Inst.47 algo1 algo4 20 34 -0.128738 0.07749
#>   Inst.47 algo2 algo3 20 29 -0.957559 0.08365
#>   Inst.47 algo2 algo4 20 34 -0.317963 0.08736
#>   Inst.47 algo3 algo4 29 34  0.639596 0.09981
#>   Inst.81 algo1 algo2 20 20  0.182913 0.05231
#>   Inst.81 algo1 algo3 20 29 -0.771428 0.07246
#>   Inst.81 algo1 algo4 20 34 -0.129880 0.07756
#>   Inst.81 algo2 algo3 20 29 -0.954341 0.08397
#>   Inst.81 algo2 algo4 20 34 -0.312793 0.08738
#>   Inst.81 algo3 algo4 29 34  0.641548 0.09979
#>   Inst.48 algo1 algo2 20 20  0.186773 0.05168
#>   Inst.48 algo1 algo3 20 29 -0.769535 0.07208
#>   Inst.48 algo1 algo4 20 34 -0.129181 0.07744
#>   Inst.48 algo2 algo3 20 29 -0.956308 0.08366
#>   Inst.48 algo2 algo4 20 34 -0.315954 0.08727
#>   Inst.48 algo3 algo4 29 34  0.640354 0.09976
#>    Inst.3 algo1 algo2 20 20  0.186587 0.05169
#>    Inst.3 algo1 algo3 20 29 -0.769627 0.07210
#>    Inst.3 algo1 algo4 20 34 -0.129215 0.07745
#>    Inst.3 algo2 algo3 20 29 -0.956214 0.08366
#>    Inst.3 algo2 algo4 20 34 -0.315802 0.08727
#>    Inst.3 algo3 algo4 29 34  0.640412 0.09976
#>   Inst.41 algo1 algo2 20 20  0.189818 0.05165
#>   Inst.41 algo1 algo3 20 29 -0.768043 0.07201
#>   Inst.41 algo1 algo4 20 34 -0.128630 0.07752
#>   Inst.41 algo2 algo3 20 29 -0.957861 0.08367
#>   Inst.41 algo2 algo4 20 34 -0.318448 0.08740
#>   Inst.41 algo3 algo4 29 34  0.639412 0.09983
#>   Inst.32 algo1 algo2 20 20  0.197344 0.04818
#>   Inst.32 algo1 algo3 20 29 -0.949174 0.06814
#>   Inst.32 algo1 algo4 20 42  0.062971 0.07610
#>   Inst.32 algo2 algo3 20 29 -1.146518 0.08093
#>   Inst.32 algo2 algo4 20 42 -0.134373 0.08610
#>   Inst.32 algo3 algo4 29 42  1.012146 0.09919
#>   Inst.42 algo1 algo2 20 20  0.199080 0.04907
#>   Inst.42 algo1 algo3 20 27 -0.929249 0.06622
#>   Inst.42 algo1 algo4 20 44  0.038975 0.07829
#>   Inst.42 algo2 algo3 20 27 -1.128329 0.07972
#>   Inst.42 algo2 algo4 20 44 -0.160106 0.08843
#>   Inst.42 algo3 algo4 27 44  0.968223 0.09930
#>   Inst.43 algo1 algo2 20 20  0.198684 0.04886
#>   Inst.43 algo1 algo3 20 27 -0.929713 0.06604
#>   Inst.43 algo1 algo4 20 44  0.039505 0.07811
#>   Inst.43 algo2 algo3 20 27 -1.128397 0.07947
#>   Inst.43 algo2 algo4 20 44 -0.159180 0.08820
#>   Inst.43 algo3 algo4 27 44  0.969218 0.09908
#>    Inst.2 algo1 algo2 20 20  0.197037 0.04805
#>    Inst.2 algo1 algo3 20 29 -0.949464 0.06800
#>    Inst.2 algo1 algo4 20 42  0.063367 0.07599
#>    Inst.2 algo2 algo3 20 29 -1.146501 0.08074
#>    Inst.2 algo2 algo4 20 42 -0.133670 0.08594
#>    Inst.2 algo3 algo4 29 42  1.012831 0.09902
#>   Inst.54 algo1 algo2 20 20  0.201022 0.05020
#>   Inst.54 algo1 algo3 20 27 -0.927276 0.06708
#>   Inst.54 algo1 algo4 20 45  0.037504 0.07756
#>   Inst.54 algo2 algo3 20 27 -1.128297 0.08089
#>   Inst.54 algo2 algo4 20 45 -0.163517 0.08819
#>   Inst.54 algo3 algo4 27 45  0.964780 0.09908
#>   Inst.49 algo1 algo2 20 20  0.196945 0.04797
#>   Inst.49 algo1 algo3 20 28 -0.952517 0.06702
#>   Inst.49 algo1 algo4 20 42  0.071439 0.07731
#>   Inst.49 algo2 algo3 20 28 -1.149462 0.07988
#>   Inst.49 algo2 algo4 20 42 -0.125506 0.08705
#>   Inst.49 algo3 algo4 28 42  1.023956 0.09943
#>   Inst.99 algo1 algo2 20 20  0.198682 0.04880
#>   Inst.99 algo1 algo3 20 29 -0.947912 0.06879
#>   Inst.99 algo1 algo4 20 42  0.061250 0.07664
#>   Inst.99 algo2 algo3 20 29 -1.146594 0.08179
#>   Inst.99 algo2 algo4 20 42 -0.137432 0.08683
#>   Inst.99 algo3 algo4 29 42  1.009162 0.09996
#>   Inst.51 algo1 algo2 20 20  0.196999 0.04804
#>   Inst.51 algo1 algo3 20 29 -0.949499 0.06799
#>   Inst.51 algo1 algo4 20 42  0.063415 0.07598
#>   Inst.51 algo2 algo3 20 29 -1.146499 0.08072
#>   Inst.51 algo2 algo4 20 42 -0.133584 0.08592
#>   Inst.51 algo3 algo4 29 42  1.012915 0.09900
#>    Inst.6 algo1 algo2 20 20  0.199822 0.04949
#>    Inst.6 algo1 algo3 20 27 -0.928378 0.06658
#>    Inst.6 algo1 algo4 20 44  0.037981 0.07863
#>    Inst.6 algo2 algo3 20 27 -1.128200 0.08021
#>    Inst.6 algo2 algo4 20 44 -0.161841 0.08889
#>    Inst.6 algo3 algo4 27 44  0.966359 0.09974
#>   Inst.77 algo1 algo2 20 20  0.197308 0.04816
#>   Inst.77 algo1 algo3 20 29 -0.949208 0.06813
#>   Inst.77 algo1 algo4 20 42  0.063018 0.07609
#>   Inst.77 algo2 algo3 20 29 -1.146516 0.08091
#>   Inst.77 algo2 algo4 20 42 -0.134290 0.08608
#>   Inst.77 algo3 algo4 29 42  1.012227 0.09917
#>   Inst.29 algo1 algo2 20 20  0.203992 0.05248
#>   Inst.29 algo1 algo3 20 31 -0.950244 0.06837
#>   Inst.29 algo1 algo4 20 44  0.056261 0.07704
#>   Inst.29 algo2 algo3 20 31 -1.154236 0.08282
#>   Inst.29 algo2 algo4 20 44 -0.147731 0.08842
#>   Inst.29 algo3 algo4 31 44  1.006505 0.09916
#>   Inst.71 algo1 algo2 20 20  0.196922 0.04796
#>   Inst.71 algo1 algo3 20 28 -0.952536 0.06701
#>   Inst.71 algo1 algo4 20 42  0.071457 0.07730
#>   Inst.71 algo2 algo3 20 28 -1.149458 0.07987
#>   Inst.71 algo2 algo4 20 42 -0.125466 0.08704
#>   Inst.71 algo3 algo4 28 42  1.023992 0.09942
#>   Inst.85 algo1 algo2 20 20  0.200489 0.04986
#>   Inst.85 algo1 algo3 20 27 -0.927950 0.06680
#>   Inst.85 algo1 algo4 20 45  0.038400 0.07729
#>   Inst.85 algo2 algo3 20 27 -1.128439 0.08052
#>   Inst.85 algo2 algo4 20 45 -0.162089 0.08784
#>   Inst.85 algo3 algo4 27 45  0.966350 0.09876
#>   Inst.57 algo1 algo2 20 20  0.197169 0.04810
#>   Inst.57 algo1 algo3 20 29 -0.949339 0.06806
#>   Inst.57 algo1 algo4 20 42  0.063197 0.07604
#>   Inst.57 algo2 algo3 20 29 -1.146508 0.08082
#>   Inst.57 algo2 algo4 20 42 -0.133973 0.08601
#>   Inst.57 algo3 algo4 29 42  1.012536 0.09909
#>    Inst.8 algo1 algo2 20 20  0.262644 0.05139
#>    Inst.8 algo1 algo3 20 32 -0.914896 0.07762
#>    Inst.8 algo1 algo4 20 32 -0.088250 0.07229
#>    Inst.8 algo2 algo3 20 32 -1.177540 0.08758
#>    Inst.8 algo2 algo4 20 32 -0.350894 0.08093
#>    Inst.8 algo3 algo4 32 32  0.826646 0.09889
#>   Inst.26 algo1 algo2 20 20  0.267751 0.05149
#>   Inst.26 algo1 algo3 20 30 -0.952517 0.07507
#>   Inst.26 algo1 algo4 20 35 -0.028864 0.07416
#>   Inst.26 algo2 algo3 20 30 -1.220268 0.08558
#>   Inst.26 algo2 algo4 20 35 -0.296616 0.08260
#>   Inst.26 algo3 algo4 30 35  0.923653 0.09878
#>   Inst.17 algo1 algo2 20 20  0.270345 0.05155
#>   Inst.17 algo1 algo3 20 30 -0.977317 0.07343
#>   Inst.17 algo1 algo4 20 34  0.004402 0.07531
#>   Inst.17 algo2 algo3 20 30 -1.247662 0.08427
#>   Inst.17 algo2 algo4 20 34 -0.265943 0.08363
#>   Inst.17 algo3 algo4 30 34  0.981720 0.09868
#>   Inst.58 algo1 algo2 20 20  0.259873 0.05173
#>   Inst.58 algo1 algo3 20 30 -0.912744 0.07487
#>   Inst.58 algo1 algo4 20 37 -0.091075 0.07596
#>   Inst.58 algo2 algo3 20 30 -1.172617 0.08512
#>   Inst.58 algo2 algo4 20 37 -0.350948 0.08421
#>   Inst.58 algo3 algo4 30 37  0.821669 0.09930
#>   Inst.91 algo1 algo2 20 20  0.270841 0.05156
#>   Inst.91 algo1 algo3 20 30 -0.976498 0.07330
#>   Inst.91 algo1 algo4 20 34  0.005151 0.07531
#>   Inst.91 algo2 algo3 20 30 -1.247339 0.08417
#>   Inst.91 algo2 algo4 20 34 -0.265690 0.08365
#>   Inst.91 algo3 algo4 30 34  0.981649 0.09860
#>   Inst.60 algo1 algo2 20 20  0.270579 0.05155
#>   Inst.60 algo1 algo3 20 30 -0.976931 0.07337
#>   Inst.60 algo1 algo4 20 34  0.004755 0.07531
#>   Inst.60 algo2 algo3 20 30 -1.247510 0.08422
#>   Inst.60 algo2 algo4 20 34 -0.265823 0.08364
#>   Inst.60 algo3 algo4 30 34  0.981686 0.09864
#>   Inst.76 algo1 algo2 20 20  0.269371 0.05152
#>   Inst.76 algo1 algo3 20 30 -0.978928 0.07368
#>   Inst.76 algo1 algo4 20 34  0.002932 0.07530
#>   Inst.76 algo2 algo3 20 30 -1.248298 0.08447
#>   Inst.76 algo2 algo4 20 34 -0.266439 0.08361
#>   Inst.76 algo3 algo4 30 34  0.981859 0.09884
#>   Inst.98 algo1 algo2 20 20  0.265007 0.05145
#>   Inst.98 algo1 algo3 20 32 -0.910709 0.07732
#>   Inst.98 algo1 algo4 20 31 -0.081263 0.07390
#>   Inst.98 algo2 algo3 20 32 -1.175716 0.08743
#>   Inst.98 algo2 algo4 20 31 -0.346269 0.08243
#>   Inst.98 algo3 algo4 32 31  0.829446 0.09991
#>   Inst.35 algo1 algo2 20 20  0.268649 0.05151
#>   Inst.35 algo1 algo3 20 30 -0.980120 0.07388
#>   Inst.35 algo1 algo4 20 34  0.001843 0.07531
#>   Inst.35 algo2 algo3 20 30 -1.248769 0.08462
#>   Inst.35 algo2 algo4 20 34 -0.266807 0.08359
#>   Inst.35 algo3 algo4 30 34  0.981963 0.09897
#>   Inst.72 algo1 algo2 20 20  0.270585 0.05155
#>   Inst.72 algo1 algo3 20 30 -0.976920 0.07337
#>   Inst.72 algo1 algo4 20 34  0.004765 0.07531
#>   Inst.72 algo2 algo3 20 30 -1.247505 0.08422
#>   Inst.72 algo2 algo4 20 34 -0.265820 0.08364
#>   Inst.72 algo3 algo4 30 34  0.981685 0.09864
#>   Inst.94 algo1 algo2 20 20  0.270750 0.05156
#>   Inst.94 algo1 algo3 20 30 -0.976647 0.07332
#>   Inst.94 algo1 algo4 20 34  0.005014 0.07531
#>   Inst.94 algo2 algo3 20 30 -1.247398 0.08419
#>   Inst.94 algo2 algo4 20 34 -0.265736 0.08365
#>   Inst.94 algo3 algo4 30 34  0.981661 0.09861
#>   Inst.73 algo1 algo2 20 20  0.270273 0.05154
#>   Inst.73 algo1 algo3 20 30 -0.977437 0.07345
#>   Inst.73 algo1 algo4 20 34  0.004293 0.07531
#>   Inst.73 algo2 algo3 20 30 -1.247710 0.08428
#>   Inst.73 algo2 algo4 20 34 -0.265980 0.08363
#>   Inst.73 algo3 algo4 30 34  0.981730 0.09869
#>   Inst.50 algo1 algo2 20 20  0.265762 0.05145
#>   Inst.50 algo1 algo3 20 32 -0.909344 0.07709
#>   Inst.50 algo1 algo4 20 31 -0.080300 0.07391
#>   Inst.50 algo2 algo3 20 32 -1.175106 0.08724
#>   Inst.50 algo2 algo4 20 31 -0.346061 0.08246
#>   Inst.50 algo3 algo4 32 31  0.829045 0.09977
#>   Inst.10 algo1 algo2 20 20  0.270829 0.05156
#>   Inst.10 algo1 algo3 20 30 -0.976518 0.07330
#>   Inst.10 algo1 algo4 20 34  0.005132 0.07531
#>   Inst.10 algo2 algo3 20 30 -1.247347 0.08418
#>   Inst.10 algo2 algo4 20 34 -0.265696 0.08365
#>   Inst.10 algo3 algo4 30 34  0.981650 0.09860
#>   Inst.45 algo1 algo2 20 20  0.270842 0.05156
#>   Inst.45 algo1 algo3 20 30 -0.976496 0.07330
#>   Inst.45 algo1 algo4 20 34  0.005153 0.07531
#>   Inst.45 algo2 algo3 20 30 -1.247338 0.08417
#>   Inst.45 algo2 algo4 20 34 -0.265689 0.08365
#>   Inst.45 algo3 algo4 30 34  0.981648 0.09860
#> 
#> #====================================