Parametric standard errors

Calculates the standard errors of a given statistic using parametric formulas

se_param(Xk, dif = "simple", comparisons = "all.vs.all", ...)

Arguments

Xk
list object where each position contains a vector of observations of algorithm k on a given problem instance.

dif
name of the difference for which the SEs are desired. Accepts "perc" (for percent differences) or "simple" (for simple differences)

comparisons
standard errors to be calculated. Accepts "all.vs.first" (in which cases the first object in algorithms is considered to be the reference algorithm) or "all.vs.all" (if there is no reference and all pairwise SEs are desired).

...
other parameters (used only for compatibility with calls to se_boot(), unused in this function)

Value

Data frame containing, for each pair of interest, the estimated difference (column "Phi") and the sample standard error (column "SE")

References

E.C. Fieller: Some problems in interval estimation. Journal of the Royal Statistical Society. Series B (Methodological) 16(2), 175–185 (1954)

V. Franz: Ratios: A short guide to confidence limits and proper use (2007). https://arxiv.org/pdf/0710.2024v1.pdf

D.C. Montgomery, C.G. Runger: Applied Statistics and Probability for Engineers, 6th ed. Wiley (2013)

F. Campelo, F. Takahashi: Sample size estimation for power and accuracy in the experimental comparison of algorithms. Journal of Heuristics 25(2):305-338, 2019.

Examples

# three vectors of normally distributed observations
set.seed(1234)
Xk <- list(rnorm(10, 5, 1),  # mean = 5, sd = 1,
           rnorm(20, 10, 2), # mean = 10, sd = 2,
           rnorm(20, 15, 5)) # mean = 15, sd = 3

se_param(Xk, dif = "simple", comparisons = "all.vs.all")

#>   Alg1 Alg2 N1 N2       Phi        SE   r      ropt
#> 1    1    2 10 20 -4.877040 0.5029276 0.5 0.5678141
#> 2    1    3 10 20 -6.943182 0.9734340 0.5 0.2417395
#> 3    2    3 20 20 -2.066142 1.0010950 1.0 0.4257371

se_param(Xk, dif = "perc", comparisons = "all.vs.first")

#>   Alg1 Alg2 N1 N2       Phi        SE   r     ropt
#> 1    1    2 10 20 -1.056358 0.1639697 0.5 1.167629
#> 2    1    3 10 20 -1.503881 0.2626191 0.5 0.605287

se_param(Xk, dif = "perc", comparisons = "all.vs.all")

#>   Alg1 Alg2 N1 N2        Phi        SE   r      ropt
#> 1    1    2 10 20 -0.5699530 0.0632245 0.5 0.5678141
#> 2    1    3 10 20 -0.8114117 0.1184975 0.5 0.2417395
#> 3    2    3 20 20 -0.2414587 0.1174082 1.0 0.4257371