print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : .
#> Sampling algorithms on instance Inst.80 : .
#> Sampling algorithms on instance Inst.22 : .
#> Sampling algorithms on instance Inst.9 : .
#> Sampling algorithms on instance Inst.5 : .
#> Sampling algorithms on instance Inst.38 : .
#> Sampling algorithms on instance Inst.16 : .
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.39693 0.05990
#>   Inst.28 algo1 algo3 20 28 -0.96605 0.07459
#>   Inst.28 algo1 algo4 20 32  0.07499 0.07624
#>   Inst.28 algo2 algo3 20 28 -1.36298 0.09028
#>   Inst.28 algo2 algo4 20 32 -0.32194 0.08805
#>   Inst.28 algo3 algo4 28 32  1.04105 0.09944
#>   Inst.80 algo1 algo2 20 20  0.39691 0.06151
#>   Inst.80 algo1 algo3 20 27 -0.95326 0.07602
#>   Inst.80 algo1 algo4 20 33  0.05589 0.07604
#>   Inst.80 algo2 algo3 20 27 -1.35016 0.09130
#>   Inst.80 algo2 algo4 20 33 -0.34102 0.08776
#>   Inst.80 algo3 algo4 27 33  1.00915 0.09908
#>   Inst.22 algo1 algo2 20 20  0.39712 0.06040
#>   Inst.22 algo1 algo3 20 28 -0.96520 0.07457
#>   Inst.22 algo1 algo4 20 32  0.07352 0.07632
#>   Inst.22 algo2 algo3 20 28 -1.36232 0.09025
#>   Inst.22 algo2 algo4 20 32 -0.32360 0.08811
#>   Inst.22 algo3 algo4 28 32  1.03872 0.09916
#>    Inst.9 algo1 algo2 20 20  0.39697 0.06000
#>    Inst.9 algo1 algo3 20 28 -0.96587 0.07458
#>    Inst.9 algo1 algo4 20 32  0.07467 0.07625
#>    Inst.9 algo2 algo3 20 28 -1.36284 0.09027
#>    Inst.9 algo2 algo4 20 32 -0.32230 0.08806
#>    Inst.9 algo3 algo4 28 32  1.04053 0.09938
#>    Inst.5 algo1 algo2 20 20  0.39715 0.06049
#>    Inst.5 algo1 algo3 20 28 -0.96506 0.07457
#>    Inst.5 algo1 algo4 20 32  0.07326 0.07634
#>    Inst.5 algo2 algo3 20 28 -1.36221 0.09024
#>    Inst.5 algo2 algo4 20 32 -0.32389 0.08812
#>    Inst.5 algo3 algo4 28 32  1.03832 0.09911
#>   Inst.38 algo1 algo2 20 20  0.39712 0.06038
#>   Inst.38 algo1 algo3 20 28 -0.96523 0.07457
#>   Inst.38 algo1 algo4 20 32  0.07355 0.07632
#>   Inst.38 algo2 algo3 20 28 -1.36234 0.09025
#>   Inst.38 algo2 algo4 20 32 -0.32356 0.08811
#>   Inst.38 algo3 algo4 28 32  1.03878 0.09917
#>   Inst.16 algo1 algo2 20 20  0.39694 0.05992
#>   Inst.16 algo1 algo3 20 28 -0.96601 0.07459
#>   Inst.16 algo1 algo4 20 32  0.07493 0.07624
#>   Inst.16 algo2 algo3 20 28 -1.36295 0.09028
#>   Inst.16 algo2 algo4 20 32 -0.32201 0.08806
#>   Inst.16 algo3 algo4 28 32  1.04094 0.09943
#>    Inst.4 algo1 algo2 20 20  0.25802 0.05371
#>    Inst.4 algo1 algo3 20 31 -0.92643 0.06890
#>    Inst.4 algo1 algo4 20 45  0.02518 0.07718
#>    Inst.4 algo2 algo3 20 31 -1.18445 0.08409
#>    Inst.4 algo2 algo4 20 45 -0.23284 0.08895
#>    Inst.4 algo3 algo4 31 45  0.95161 0.09909
#>   Inst.86 algo1 algo2 20 20  0.25571 0.05303
#>   Inst.86 algo1 algo3 20 31 -0.92117 0.06840
#>   Inst.86 algo1 algo4 20 45  0.01112 0.07822
#>   Inst.86 algo2 algo3 20 31 -1.17687 0.08341
#>   Inst.86 algo2 algo4 20 45 -0.24459 0.08966
#>   Inst.86 algo3 algo4 31 45  0.93228 0.09962
#>   Inst.90 algo1 algo2 20 20  0.25627 0.05317
#>   Inst.90 algo1 algo3 20 30 -0.89560 0.06879
#>   Inst.90 algo1 algo4 20 46 -0.01145 0.07775
#>   Inst.90 algo2 algo3 20 30 -1.15188 0.08374
#>   Inst.90 algo2 algo4 20 46 -0.26772 0.08936
#>   Inst.90 algo3 algo4 30 46  0.88415 0.09935
#>   Inst.70 algo1 algo2 20 20  0.25735 0.05350
#>   Inst.70 algo1 algo3 20 31 -0.92840 0.06885
#>   Inst.70 algo1 algo4 20 45  0.02451 0.07729
#>   Inst.70 algo2 algo3 20 31 -1.18575 0.08397
#>   Inst.70 algo2 algo4 20 45 -0.23284 0.08898
#>   Inst.70 algo3 algo4 31 45  0.95291 0.09919
#>   Inst.79 algo1 algo2 20 20  0.25432 0.05269
#>   Inst.79 algo1 algo3 20 32 -0.89861 0.06949
#>   Inst.79 algo1 algo4 20 45 -0.01975 0.07677
#>   Inst.79 algo2 algo3 20 32 -1.15292 0.08410
#>   Inst.79 algo2 algo4 20 45 -0.27407 0.08833
#>   Inst.79 algo3 algo4 32 45  0.87886 0.09910
#>   Inst.78 algo1 algo2 20 20  0.25627 0.05318
#>   Inst.78 algo1 algo3 20 31 -0.93157 0.06879
#>   Inst.78 algo1 algo4 20 45  0.02342 0.07748
#>   Inst.78 algo2 algo3 20 31 -1.18783 0.08381
#>   Inst.78 algo2 algo4 20 45 -0.23285 0.08904
#>   Inst.78 algo3 algo4 31 45  0.95498 0.09937
#>   Inst.14 algo1 algo2 20 20  0.25737 0.05350
#>   Inst.14 algo1 algo3 20 31 -0.92834 0.06885
#>   Inst.14 algo1 algo4 20 45  0.02453 0.07728
#>   Inst.14 algo2 algo3 20 31 -1.18571 0.08397
#>   Inst.14 algo2 algo4 20 45 -0.23284 0.08898
#>   Inst.14 algo3 algo4 31 45  0.95286 0.09919
#>   Inst.56 algo1 algo2 20 20  0.25693 0.05337
#>   Inst.56 algo1 algo3 20 31 -0.92962 0.06882
#>   Inst.56 algo1 algo4 20 45  0.02409 0.07736
#>   Inst.56 algo2 algo3 20 31 -1.18655 0.08390
#>   Inst.56 algo2 algo4 20 45 -0.23284 0.08900
#>   Inst.56 algo3 algo4 31 45  0.95370 0.09926
#>   Inst.62 algo1 algo2 20 20  0.25618 0.05316
#>   Inst.62 algo1 algo3 20 31 -0.93181 0.06879
#>   Inst.62 algo1 algo4 20 45  0.02333 0.07749
#>   Inst.62 algo2 algo3 20 31 -1.18799 0.08379
#>   Inst.62 algo2 algo4 20 45 -0.23285 0.08905
#>   Inst.62 algo3 algo4 31 45  0.95514 0.09938
#>   Inst.93 algo1 algo2 20 20  0.25790 0.05367
#>   Inst.93 algo1 algo3 20 31 -0.92679 0.06889
#>   Inst.93 algo1 algo4 20 45  0.02506 0.07719
#>   Inst.93 algo2 algo3 20 31 -1.18468 0.08407
#>   Inst.93 algo2 algo4 20 45 -0.23284 0.08896
#>   Inst.93 algo3 algo4 31 45  0.95185 0.09911
#> 
#> #====================================