print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE, breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.39267 0.05793
#>   Inst.28 algo1 algo3 20 32 -0.93802 0.07267
#>   Inst.28 algo1 algo4 20 39  0.02885 0.07312
#>   Inst.28 algo2 algo3 20 32 -1.33070 0.09081
#>   Inst.28 algo2 algo4 20 39 -0.36383 0.08793
#>   Inst.28 algo3 algo4 32 39  0.96687 0.09856
#>   Inst.80 algo1 algo2 20 20  0.39920 0.05777
#>   Inst.80 algo1 algo3 20 32 -0.94592 0.07160
#>   Inst.80 algo1 algo4 20 39  0.04080 0.07527
#>   Inst.80 algo2 algo3 20 32 -1.34512 0.08951
#>   Inst.80 algo2 algo4 20 39 -0.35840 0.08914
#>   Inst.80 algo3 algo4 32 39  0.98672 0.09909
#>   Inst.22 algo1 algo2 20 20  0.39500 0.05784
#>   Inst.22 algo1 algo3 20 32 -0.94914 0.07107
#>   Inst.22 algo1 algo4 20 39  0.04198 0.07508
#>   Inst.22 algo2 algo3 20 32 -1.34414 0.08943
#>   Inst.22 algo2 algo4 20 39 -0.35302 0.08935
#>   Inst.22 algo3 algo4 32 39  0.99112 0.09888
#>    Inst.9 algo1 algo2 20 20  0.39320 0.05791
#>    Inst.9 algo1 algo3 20 32 -0.94293 0.07139
#>    Inst.9 algo1 algo4 20 39  0.03429 0.07462
#>    Inst.9 algo2 algo3 20 32 -1.33613 0.08979
#>    Inst.9 algo2 algo4 20 39 -0.35891 0.08913
#>    Inst.9 algo3 algo4 32 39  0.97722 0.09878
#>    Inst.5 algo1 algo2 20 20  0.39538 0.05782
#>    Inst.5 algo1 algo3 20 32 -0.94898 0.07106
#>    Inst.5 algo1 algo4 20 39  0.04203 0.07513
#>    Inst.5 algo2 algo3 20 32 -1.34436 0.08940
#>    Inst.5 algo2 algo4 20 39 -0.35336 0.08937
#>    Inst.5 algo3 algo4 32 39  0.99100 0.09889
#>   Inst.38 algo1 algo2 20 20  0.39494 0.05784
#>   Inst.38 algo1 algo3 20 32 -0.94916 0.07107
#>   Inst.38 algo1 algo4 20 39  0.04198 0.07507
#>   Inst.38 algo2 algo3 20 32 -1.34410 0.08944
#>   Inst.38 algo2 algo4 20 39 -0.35297 0.08935
#>   Inst.38 algo3 algo4 32 39  0.99114 0.09888
#>   Inst.16 algo1 algo2 20 20  0.39281 0.05794
#>   Inst.16 algo1 algo3 20 32 -0.94306 0.07140
#>   Inst.16 algo1 algo4 20 39  0.03422 0.07458
#>   Inst.16 algo2 algo3 20 32 -1.33587 0.08982
#>   Inst.16 algo2 algo4 20 39 -0.35859 0.08913
#>   Inst.16 algo3 algo4 32 39  0.97727 0.09877
#>    Inst.4 algo1 algo2 20 20  0.39233 0.05802
#>    Inst.4 algo1 algo3 20 32 -0.93873 0.07282
#>    Inst.4 algo1 algo4 20 38  0.03077 0.07485
#>    Inst.4 algo2 algo3 20 32 -1.33106 0.09107
#>    Inst.4 algo2 algo4 20 38 -0.36156 0.08943
#>    Inst.4 algo3 algo4 32 38  0.96950 0.10000
#>   Inst.86 algo1 algo2 20 20  0.39653 0.05775
#>   Inst.86 algo1 algo3 20 31 -0.92231 0.07212
#>   Inst.86 algo1 algo4 20 40  0.01408 0.07440
#>   Inst.86 algo2 algo3 20 31 -1.31884 0.09004
#>   Inst.86 algo2 algo4 20 40 -0.38246 0.08871
#>   Inst.86 algo3 algo4 31 40  0.93638 0.09882
#>   Inst.90 algo1 algo2 20 20  0.39588 0.05781
#>   Inst.90 algo1 algo3 20 32 -0.94877 0.07106
#>   Inst.90 algo1 algo4 20 39  0.04209 0.07520
#>   Inst.90 algo2 algo3 20 32 -1.34465 0.08935
#>   Inst.90 algo2 algo4 20 39 -0.35379 0.08938
#>   Inst.90 algo3 algo4 32 39  0.99085 0.09891
#>   Inst.70 algo1 algo2 20 20  0.39347 0.05790
#>   Inst.70 algo1 algo3 20 32 -0.94284 0.07139
#>   Inst.70 algo1 algo4 20 39  0.03435 0.07465
#>   Inst.70 algo2 algo3 20 32 -1.33631 0.08977
#>   Inst.70 algo2 algo4 20 39 -0.35913 0.08914
#>   Inst.70 algo3 algo4 32 39  0.97718 0.09879
#>   Inst.79 algo1 algo2 20 20  0.39879 0.05777
#>   Inst.79 algo1 algo3 20 32 -0.94606 0.07157
#>   Inst.79 algo1 algo4 20 39  0.04073 0.07522
#>   Inst.79 algo2 algo3 20 32 -1.34485 0.08952
#>   Inst.79 algo2 algo4 20 39 -0.35807 0.08914
#>   Inst.79 algo3 algo4 32 39  0.98679 0.09907
#>   Inst.78 algo1 algo2 20 20  0.39564 0.05781
#>   Inst.78 algo1 algo3 20 32 -0.94887 0.07106
#>   Inst.78 algo1 algo4 20 39  0.04206 0.07517
#>   Inst.78 algo2 algo3 20 32 -1.34451 0.08938
#>   Inst.78 algo2 algo4 20 39 -0.35358 0.08938
#>   Inst.78 algo3 algo4 32 39  0.99093 0.09890
#>   Inst.14 algo1 algo2 20 20  0.39343 0.05790
#>   Inst.14 algo1 algo3 20 32 -0.94285 0.07139
#>   Inst.14 algo1 algo4 20 39  0.03434 0.07465
#>   Inst.14 algo2 algo3 20 32 -1.33629 0.08977
#>   Inst.14 algo2 algo4 20 39 -0.35910 0.08914
#>   Inst.14 algo3 algo4 32 39  0.97719 0.09879
#>   Inst.56 algo1 algo2 20 20  0.39435 0.05786
#>   Inst.56 algo1 algo3 20 32 -0.94941 0.07109
#>   Inst.56 algo1 algo4 20 39  0.04191 0.07499
#>   Inst.56 algo2 algo3 20 32 -1.34376 0.08949
#>   Inst.56 algo2 algo4 20 39 -0.35244 0.08933
#>   Inst.56 algo3 algo4 32 39  0.99132 0.09885
#>   Inst.62 algo1 algo2 20 20  0.39580 0.05781
#>   Inst.62 algo1 algo3 20 32 -0.94880 0.07106
#>   Inst.62 algo1 algo4 20 39  0.04208 0.07519
#>   Inst.62 algo2 algo3 20 32 -1.34460 0.08936
#>   Inst.62 algo2 algo4 20 39 -0.35372 0.08938
#>   Inst.62 algo3 algo4 32 39  0.99088 0.09891
#>   Inst.93 algo1 algo2 20 20  0.39237 0.05795
#>   Inst.93 algo1 algo3 20 32 -0.93816 0.07268
#>   Inst.93 algo1 algo4 20 39  0.02882 0.07309
#>   Inst.93 algo2 algo3 20 32 -1.33054 0.09084
#>   Inst.93 algo2 algo4 20 39 -0.36355 0.08793
#>   Inst.93 algo3 algo4 32 39  0.96699 0.09856
#> 
#> #====================================