print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : .
#> Sampling algorithms on instance Inst.38 : .
#> Sampling algorithms on instance Inst.16 : .
#> Sampling algorithms on instance Inst.4 : .
#> Sampling algorithms on instance Inst.86 : .
#> Sampling algorithms on instance Inst.90 : .
#> Sampling algorithms on instance Inst.70 : .
#> Sampling algorithms on instance Inst.79 : .
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : .
#> Sampling algorithms on instance Inst.93 : .

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.34186 0.05688
#>   Inst.28 algo1 algo3 20 32 -0.78483 0.07490
#>   Inst.28 algo1 algo4 20 40  0.04396 0.07688
#>   Inst.28 algo2 algo3 20 32 -1.12670 0.08742
#>   Inst.28 algo2 algo4 20 40 -0.29790 0.08659
#>   Inst.28 algo3 algo4 32 40  0.82880 0.09975
#>   Inst.80 algo1 algo2 20 20  0.33468 0.05815
#>   Inst.80 algo1 algo3 20 28 -0.80193 0.07143
#>   Inst.80 algo1 algo4 20 42  0.04646 0.08041
#>   Inst.80 algo2 algo3 20 28 -1.13661 0.08497
#>   Inst.80 algo2 algo4 20 42 -0.28822 0.09018
#>   Inst.80 algo3 algo4 28 42  0.84839 0.09969
#>   Inst.22 algo1 algo2 20 20  0.34013 0.05736
#>   Inst.22 algo1 algo3 20 29 -0.79258 0.07197
#>   Inst.22 algo1 algo4 20 43  0.05384 0.07930
#>   Inst.22 algo2 algo3 20 29 -1.13271 0.08513
#>   Inst.22 algo2 algo4 20 43 -0.28629 0.08890
#>   Inst.22 algo3 algo4 29 43  0.84642 0.09945
#>    Inst.9 algo1 algo2 20 20  0.34141 0.05697
#>    Inst.9 algo1 algo3 20 32 -0.78514 0.07479
#>    Inst.9 algo1 algo4 20 40  0.04405 0.07687
#>    Inst.9 algo2 algo3 20 32 -1.12655 0.08735
#>    Inst.9 algo2 algo4 20 40 -0.29735 0.08662
#>    Inst.9 algo3 algo4 32 40  0.82920 0.09964
#>    Inst.5 algo1 algo2 20 20  0.22126 0.07090
#>    Inst.5 algo1 algo3 20 24 -0.87061 0.07132
#>    Inst.5 algo1 algo4 20 35  0.24752 0.07765
#>    Inst.5 algo2 algo3 20 24 -1.09187 0.09420
#>    Inst.5 algo2 algo4 20 35  0.02626 0.09690
#>    Inst.5 algo3 algo4 24 35  1.11813 0.09958
#>   Inst.38 algo1 algo2 20 20  0.22096 0.07084
#>   Inst.38 algo1 algo3 20 24 -0.87044 0.07125
#>   Inst.38 algo1 algo4 20 35  0.24747 0.07756
#>   Inst.38 algo2 algo3 20 24 -1.09140 0.09417
#>   Inst.38 algo2 algo4 20 35  0.02651 0.09685
#>   Inst.38 algo3 algo4 24 35  1.11791 0.09952
#>   Inst.16 algo1 algo2 20 20  0.21958 0.07059
#>   Inst.16 algo1 algo3 20 24 -0.86965 0.07097
#>   Inst.16 algo1 algo4 20 35  0.24727 0.07720
#>   Inst.16 algo2 algo3 20 24 -1.08922 0.09401
#>   Inst.16 algo2 algo4 20 35  0.02769 0.09665
#>   Inst.16 algo3 algo4 24 35  1.11691 0.09928
#>    Inst.4 algo1 algo2 20 20  0.21915 0.07051
#>    Inst.4 algo1 algo3 20 24 -0.86940 0.07089
#>    Inst.4 algo1 algo4 20 35  0.24721 0.07709
#>    Inst.4 algo2 algo3 20 24 -1.08855 0.09397
#>    Inst.4 algo2 algo4 20 35  0.02806 0.09659
#>    Inst.4 algo3 algo4 24 35  1.11661 0.09921
#>   Inst.86 algo1 algo2 20 20  0.22220 0.07109
#>   Inst.86 algo1 algo3 20 24 -0.87115 0.07154
#>   Inst.86 algo1 algo4 20 35  0.24766 0.07793
#>   Inst.86 algo2 algo3 20 24 -1.09336 0.09433
#>   Inst.86 algo2 algo4 20 35  0.02545 0.09706
#>   Inst.86 algo3 algo4 24 35  1.11881 0.09977
#>   Inst.90 algo1 algo2 20 20  0.22159 0.07097
#>   Inst.90 algo1 algo3 20 24 -0.87080 0.07139
#>   Inst.90 algo1 algo4 20 35  0.24757 0.07774
#>   Inst.90 algo2 algo3 20 24 -1.09239 0.09424
#>   Inst.90 algo2 algo4 20 35  0.02598 0.09695
#>   Inst.90 algo3 algo4 24 35  1.11837 0.09964
#>   Inst.70 algo1 algo2 20 20  0.22002 0.07067
#>   Inst.70 algo1 algo3 20 24 -0.86990 0.07105
#>   Inst.70 algo1 algo4 20 35  0.24734 0.07731
#>   Inst.70 algo2 algo3 20 24 -1.08992 0.09406
#>   Inst.70 algo2 algo4 20 35  0.02731 0.09671
#>   Inst.70 algo3 algo4 24 35  1.11724 0.09935
#>   Inst.79 algo1 algo2 20 20  0.22350 0.07137
#>   Inst.79 algo1 algo3 20 25 -0.87275 0.06973
#>   Inst.79 algo1 algo4 20 35  0.24779 0.07833
#>   Inst.79 algo2 algo3 20 25 -1.09625 0.09283
#>   Inst.79 algo2 algo4 20 35  0.02429 0.09729
#>   Inst.79 algo3 algo4 25 35  1.12054 0.09847
#>   Inst.78 algo1 algo2 20 20  0.22143 0.07093
#>   Inst.78 algo1 algo3 20 24 -0.87071 0.07136
#>   Inst.78 algo1 algo4 20 35  0.24754 0.07769
#>   Inst.78 algo2 algo3 20 24 -1.09213 0.09422
#>   Inst.78 algo2 algo4 20 35  0.02612 0.09692
#>   Inst.78 algo3 algo4 24 35  1.11825 0.09961
#>   Inst.14 algo1 algo2 20 20  0.22000 0.07066
#>   Inst.14 algo1 algo3 20 24 -0.86989 0.07105
#>   Inst.14 algo1 algo4 20 35  0.24733 0.07730
#>   Inst.14 algo2 algo3 20 24 -1.08988 0.09406
#>   Inst.14 algo2 algo4 20 35  0.02734 0.09670
#>   Inst.14 algo3 algo4 24 35  1.11722 0.09935
#>   Inst.56 algo1 algo2 20 20  0.22056 0.07076
#>   Inst.56 algo1 algo3 20 24 -0.87021 0.07117
#>   Inst.56 algo1 algo4 20 35  0.24741 0.07745
#>   Inst.56 algo2 algo3 20 24 -1.09077 0.09412
#>   Inst.56 algo2 algo4 20 35  0.02685 0.09679
#>   Inst.56 algo3 algo4 24 35  1.11763 0.09945
#>   Inst.62 algo1 algo2 20 20  0.22153 0.07096
#>   Inst.62 algo1 algo3 20 24 -0.87077 0.07138
#>   Inst.62 algo1 algo4 20 35  0.24756 0.07773
#>   Inst.62 algo2 algo3 20 24 -1.09230 0.09424
#>   Inst.62 algo2 algo4 20 35  0.02602 0.09694
#>   Inst.62 algo3 algo4 24 35  1.11833 0.09963
#>   Inst.93 algo1 algo2 20 20  0.21931 0.07054
#>   Inst.93 algo1 algo3 20 24 -0.86949 0.07092
#>   Inst.93 algo1 algo4 20 35  0.24723 0.07713
#>   Inst.93 algo2 algo3 20 24 -1.08880 0.09399
#>   Inst.93 algo2 algo4 20 35  0.02792 0.09661
#>   Inst.93 algo3 algo4 24 35  1.11672 0.09924
#> 
#> #====================================