print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : .
#> Sampling algorithms on instance Inst.79 : .
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : .
#> Sampling algorithms on instance Inst.93 : .

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2        Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.3223077 0.05843
#>   Inst.28 algo1 algo3 20 21 -0.8642397 0.06681
#>   Inst.28 algo1 algo4 20 43 -0.0289209 0.08435
#>   Inst.28 algo2 algo3 20 21 -1.1865473 0.08119
#>   Inst.28 algo2 algo4 20 43 -0.3512286 0.09391
#>   Inst.28 algo3 algo4 21 43  0.8353187 0.09907
#>   Inst.80 algo1 algo2 20 20  0.3321455 0.05854
#>   Inst.80 algo1 algo3 20 21 -0.8736200 0.06766
#>   Inst.80 algo1 algo4 20 44 -0.0243717 0.08378
#>   Inst.80 algo2 algo3 20 21 -1.2057655 0.08216
#>   Inst.80 algo2 algo4 20 44 -0.3565172 0.09352
#>   Inst.80 algo3 algo4 21 44  0.8492482 0.09925
#>   Inst.22 algo1 algo2 20 20  0.3254687 0.05839
#>   Inst.22 algo1 algo3 20 21 -0.8668679 0.06701
#>   Inst.22 algo1 algo4 20 43 -0.0301072 0.08452
#>   Inst.22 algo2 algo3 20 21 -1.1923366 0.08145
#>   Inst.22 algo2 algo4 20 43 -0.3555759 0.09411
#>   Inst.22 algo3 algo4 21 43  0.8367607 0.09940
#>    Inst.9 algo1 algo2 20 20  0.3230037 0.05842
#>    Inst.9 algo1 algo3 20 21 -0.8648184 0.06685
#>    Inst.9 algo1 algo4 20 43 -0.0291821 0.08439
#>    Inst.9 algo2 algo3 20 21 -1.1878221 0.08124
#>    Inst.9 algo2 algo4 20 43 -0.3521858 0.09395
#>    Inst.9 algo3 algo4 21 43  0.8356362 0.09914
#>    Inst.5 algo1 algo2 20 20  0.3260112 0.05839
#>    Inst.5 algo1 algo3 20 21 -0.8673189 0.06705
#>    Inst.5 algo1 algo4 20 43 -0.0303107 0.08456
#>    Inst.5 algo2 algo3 20 21 -1.1933302 0.08150
#>    Inst.5 algo2 algo4 20 43 -0.3563220 0.09415
#>    Inst.5 algo3 algo4 21 43  0.8370082 0.09947
#>   Inst.38 algo1 algo2 20 20  0.3253868 0.05839
#>   Inst.38 algo1 algo3 20 21 -0.8667998 0.06700
#>   Inst.38 algo1 algo4 20 43 -0.0300764 0.08452
#>   Inst.38 algo2 algo3 20 21 -1.1921866 0.08144
#>   Inst.38 algo2 algo4 20 43 -0.3554633 0.09411
#>   Inst.38 algo3 algo4 21 43  0.8367234 0.09939
#>   Inst.16 algo1 algo2 20 20  0.3224488 0.05842
#>   Inst.16 algo1 algo3 20 21 -0.8643570 0.06682
#>   Inst.16 algo1 algo4 20 43 -0.0289739 0.08436
#>   Inst.16 algo2 algo3 20 21 -1.1868058 0.08120
#>   Inst.16 algo2 algo4 20 43 -0.3514227 0.09392
#>   Inst.16 algo3 algo4 21 43  0.8353831 0.09908
#>    Inst.4 algo1 algo2 20 20  0.3215405 0.05844
#>    Inst.4 algo1 algo3 20 21 -0.8636018 0.06677
#>    Inst.4 algo1 algo4 20 43 -0.0286331 0.08431
#>    Inst.4 algo2 algo3 20 21 -1.1851424 0.08113
#>    Inst.4 algo2 algo4 20 43 -0.3501736 0.09386
#>    Inst.4 algo3 algo4 21 43  0.8349688 0.09899
#>   Inst.86 algo1 algo2 20 20  0.3280193 0.05839
#>   Inst.86 algo1 algo3 20 21 -0.8689885 0.06722
#>   Inst.86 algo1 algo4 20 43 -0.0310643 0.08469
#>   Inst.86 algo2 algo3 20 21 -1.1970079 0.08170
#>   Inst.86 algo2 algo4 20 43 -0.3590836 0.09429
#>   Inst.86 algo3 algo4 21 43  0.8379242 0.09971
#>   Inst.90 algo1 algo2 20 20  0.3267153 0.05839
#>   Inst.90 algo1 algo3 20 21 -0.8679043 0.06711
#>   Inst.90 algo1 algo4 20 43 -0.0305750 0.08460
#>   Inst.90 algo2 algo3 20 21 -1.1946196 0.08157
#>   Inst.90 algo2 algo4 20 43 -0.3572903 0.09420
#>   Inst.90 algo3 algo4 21 43  0.8373294 0.09955
#>   Inst.70 algo1 algo2 20 20  0.2566742 0.03919
#>   Inst.70 algo1 algo3 20 21 -0.7857261 0.06881
#>   Inst.70 algo1 algo4 20 25  0.2576753 0.07502
#>   Inst.70 algo2 algo3 20 21 -1.0424003 0.07521
#>   Inst.70 algo2 algo4 20 25  0.0010011 0.07872
#>   Inst.70 algo3 algo4 21 25  1.0434014 0.09874
#>   Inst.79 algo1 algo2 20 20  0.2488373 0.03957
#>   Inst.79 algo1 algo3 20 21 -0.7949311 0.06927
#>   Inst.79 algo1 algo4 20 25  0.2572416 0.07568
#>   Inst.79 algo2 algo3 20 21 -1.0437684 0.07542
#>   Inst.79 algo2 algo4 20 25  0.0084043 0.07918
#>   Inst.79 algo3 algo4 21 25  1.0521727 0.09930
#>   Inst.78 algo1 algo2 20 20  0.2535686 0.03928
#>   Inst.78 algo1 algo3 20 21 -0.7893738 0.06896
#>   Inst.78 algo1 algo4 20 25  0.2575034 0.07527
#>   Inst.78 algo2 algo3 20 21 -1.0429424 0.07524
#>   Inst.78 algo2 algo4 20 25  0.0039348 0.07888
#>   Inst.78 algo3 algo4 21 25  1.0468772 0.09894
#>   Inst.14 algo1 algo2 20 20  0.2567348 0.03919
#>   Inst.14 algo1 algo3 20 21 -0.7856549 0.06881
#>   Inst.14 algo1 algo4 20 25  0.2576787 0.07502
#>   Inst.14 algo2 algo3 20 21 -1.0423897 0.07521
#>   Inst.14 algo2 algo4 20 25  0.0009438 0.07872
#>   Inst.14 algo3 algo4 21 25  1.0433335 0.09874
#>   Inst.56 algo1 algo2 20 20  0.2554808 0.03922
#>   Inst.56 algo1 algo3 20 21 -0.7871279 0.06886
#>   Inst.56 algo1 algo4 20 25  0.2576093 0.07512
#>   Inst.56 algo2 algo3 20 21 -1.0426086 0.07522
#>   Inst.56 algo2 algo4 20 25  0.0021285 0.07878
#>   Inst.56 algo3 algo4 21 25  1.0447371 0.09881
#>   Inst.62 algo1 algo2 20 20  0.2533295 0.03929
#>   Inst.62 algo1 algo3 20 21 -0.7896547 0.06897
#>   Inst.62 algo1 algo4 20 25  0.2574902 0.07529
#>   Inst.62 algo2 algo3 20 21 -1.0429842 0.07525
#>   Inst.62 algo2 algo4 20 25  0.0041607 0.07889
#>   Inst.62 algo3 algo4 21 25  1.0471449 0.09895
#>   Inst.93 algo1 algo2 20 20  0.2582583 0.03917
#>   Inst.93 algo1 algo3 20 21 -0.7838655 0.06875
#>   Inst.93 algo1 algo4 20 25  0.2577630 0.07490
#>   Inst.93 algo2 algo3 20 21 -1.0421237 0.07522
#>   Inst.93 algo2 algo4 20 25 -0.0004953 0.07866
#>   Inst.93 algo3 algo4 21 25  1.0416285 0.09866
#> 
#> #====================================