print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ...
#> Sampling algorithms on instance Inst.80 : ...
#> Sampling algorithms on instance Inst.22 : ...
#> Sampling algorithms on instance Inst.9 : ...
#> Sampling algorithms on instance Inst.5 : ...
#> Sampling algorithms on instance Inst.38 : ...
#> Sampling algorithms on instance Inst.16 : ...
#> Sampling algorithms on instance Inst.4 : ...
#> Sampling algorithms on instance Inst.86 : ...
#> Sampling algorithms on instance Inst.90 : ...
#> Sampling algorithms on instance Inst.70 : ...
#> Sampling algorithms on instance Inst.79 : ...
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : .
#> Sampling algorithms on instance Inst.93 : .

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.38208 0.05811
#>   Inst.28 algo1 algo3 20 33 -0.98909 0.07243
#>   Inst.28 algo1 algo4 20 48  0.10293 0.07798
#>   Inst.28 algo2 algo3 20 33 -1.37117 0.08697
#>   Inst.28 algo2 algo4 20 48 -0.27915 0.08810
#>   Inst.28 algo3 algo4 33 48  1.09202 0.09923
#>   Inst.80 algo1 algo2 20 20  0.38348 0.05884
#>   Inst.80 algo1 algo3 20 35 -0.96725 0.07257
#>   Inst.80 algo1 algo4 20 50  0.08144 0.07787
#>   Inst.80 algo2 algo3 20 35 -1.35072 0.08754
#>   Inst.80 algo2 algo4 20 50 -0.30204 0.08855
#>   Inst.80 algo3 algo4 35 50  1.04868 0.09907
#>   Inst.22 algo1 algo2 20 20  0.38217 0.05824
#>   Inst.22 algo1 algo3 20 33 -0.99147 0.07250
#>   Inst.22 algo1 algo4 20 48  0.10226 0.07802
#>   Inst.22 algo2 algo3 20 33 -1.37364 0.08720
#>   Inst.22 algo2 algo4 20 48 -0.27991 0.08829
#>   Inst.22 algo3 algo4 33 48  1.09373 0.09937
#>    Inst.9 algo1 algo2 20 20  0.38210 0.05813
#>    Inst.9 algo1 algo3 20 33 -0.98962 0.07244
#>    Inst.9 algo1 algo4 20 48  0.10278 0.07798
#>    Inst.9 algo2 algo3 20 33 -1.37171 0.08702
#>    Inst.9 algo2 algo4 20 48 -0.27932 0.08814
#>    Inst.9 algo3 algo4 33 48  1.09240 0.09926
#>    Inst.5 algo1 algo2 20 20  0.38340 0.05846
#>    Inst.5 algo1 algo3 20 35 -0.94520 0.07158
#>    Inst.5 algo1 algo4 20 50  0.06536 0.07912
#>    Inst.5 algo2 algo3 20 35 -1.32860 0.08642
#>    Inst.5 algo2 algo4 20 50 -0.31804 0.08948
#>    Inst.5 algo3 algo4 35 50  1.01056 0.09923
#>   Inst.38 algo1 algo2 20 20  0.38217 0.05824
#>   Inst.38 algo1 algo3 20 33 -0.99141 0.07250
#>   Inst.38 algo1 algo4 20 48  0.10228 0.07801
#>   Inst.38 algo2 algo3 20 33 -1.37358 0.08719
#>   Inst.38 algo2 algo4 20 48 -0.27989 0.08828
#>   Inst.38 algo3 algo4 33 48  1.09369 0.09936
#>   Inst.16 algo1 algo2 20 20  0.38208 0.05811
#>   Inst.16 algo1 algo3 20 33 -0.98920 0.07243
#>   Inst.16 algo1 algo4 20 48  0.10290 0.07798
#>   Inst.16 algo2 algo3 20 33 -1.37128 0.08698
#>   Inst.16 algo2 algo4 20 48 -0.27918 0.08811
#>   Inst.16 algo3 algo4 33 48  1.09210 0.09924
#>    Inst.4 algo1 algo2 20 20  0.38206 0.05809
#>    Inst.4 algo1 algo3 20 33 -0.98851 0.07242
#>    Inst.4 algo1 algo4 20 48  0.10309 0.07798
#>    Inst.4 algo2 algo3 20 33 -1.37057 0.08692
#>    Inst.4 algo2 algo4 20 48 -0.27896 0.08806
#>    Inst.4 algo3 algo4 33 48  1.09161 0.09920
#>   Inst.86 algo1 algo2 20 20  0.38344 0.05858
#>   Inst.86 algo1 algo3 20 35 -0.94649 0.07158
#>   Inst.86 algo1 algo4 20 50  0.06457 0.07915
#>   Inst.86 algo2 algo3 20 35 -1.32993 0.08652
#>   Inst.86 algo2 algo4 20 50 -0.31887 0.08959
#>   Inst.86 algo3 algo4 35 50  1.01106 0.09926
#>   Inst.90 algo1 algo2 20 20  0.38342 0.05850
#>   Inst.90 algo1 algo3 20 35 -0.94565 0.07158
#>   Inst.90 algo1 algo4 20 50  0.06509 0.07913
#>   Inst.90 algo2 algo3 20 35 -1.32907 0.08646
#>   Inst.90 algo2 algo4 20 50 -0.31833 0.08952
#>   Inst.90 algo3 algo4 35 50  1.01074 0.09924
#>   Inst.70 algo1 algo2 20 20  0.38211 0.05815
#>   Inst.70 algo1 algo3 20 33 -0.98991 0.07245
#>   Inst.70 algo1 algo4 20 48  0.10270 0.07799
#>   Inst.70 algo2 algo3 20 33 -1.37202 0.08704
#>   Inst.70 algo2 algo4 20 48 -0.27941 0.08816
#>   Inst.70 algo3 algo4 33 48  1.09261 0.09927
#>   Inst.79 algo1 algo2 20 20  0.38346 0.05878
#>   Inst.79 algo1 algo3 20 35 -0.96702 0.07256
#>   Inst.79 algo1 algo4 20 50  0.08177 0.07785
#>   Inst.79 algo2 algo3 20 35 -1.35048 0.08750
#>   Inst.79 algo2 algo4 20 50 -0.30169 0.08851
#>   Inst.79 algo3 algo4 35 50  1.04879 0.09905
#>   Inst.78 algo1 algo2 20 20  0.30660 0.06702
#>   Inst.78 algo1 algo3 20 25 -0.91882 0.07671
#>   Inst.78 algo1 algo4 20 30  0.19952 0.07521
#>   Inst.78 algo2 algo3 20 25 -1.22542 0.09476
#>   Inst.78 algo2 algo4 20 30 -0.10709 0.09044
#>   Inst.78 algo3 algo4 25 30  1.11833 0.09990
#>   Inst.14 algo1 algo2 20 20  0.30534 0.06560
#>   Inst.14 algo1 algo3 20 25 -0.92147 0.07612
#>   Inst.14 algo1 algo4 20 30  0.19544 0.07468
#>   Inst.14 algo2 algo3 20 25 -1.22681 0.09403
#>   Inst.14 algo2 algo4 20 30 -0.10990 0.08981
#>   Inst.14 algo3 algo4 25 30  1.11691 0.09975
#>   Inst.56 algo1 algo2 20 20  0.30584 0.06614
#>   Inst.56 algo1 algo3 20 25 -0.92042 0.07634
#>   Inst.56 algo1 algo4 20 30  0.19706 0.07488
#>   Inst.56 algo2 algo3 20 25 -1.22626 0.09431
#>   Inst.56 algo2 algo4 20 30 -0.10879 0.09005
#>   Inst.56 algo3 algo4 25 30  1.11747 0.09980
#>   Inst.62 algo1 algo2 20 20  0.30670 0.06713
#>   Inst.62 algo1 algo3 20 25 -0.91862 0.07676
#>   Inst.62 algo1 algo4 20 30  0.19982 0.07526
#>   Inst.62 algo2 algo3 20 25 -1.22532 0.09482
#>   Inst.62 algo2 algo4 20 30 -0.10688 0.09050
#>   Inst.62 algo3 algo4 25 30  1.11844 0.09992
#>   Inst.93 algo1 algo2 20 20  0.30473 0.06499
#>   Inst.93 algo1 algo3 20 25 -0.92092 0.07586
#>   Inst.93 algo1 algo4 20 30  0.19155 0.07465
#>   Inst.93 algo2 algo3 20 25 -1.22565 0.09371
#>   Inst.93 algo2 algo4 20 30 -0.11317 0.08968
#>   Inst.93 algo3 algo4 25 30  1.11247 0.09980
#> 
#> #====================================