print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.19762 0.04330
#>   Inst.28 algo1 algo3 20 34 -0.84995 0.07599
#>   Inst.28 algo1 algo4 20 37 -0.02741 0.07343
#>   Inst.28 algo2 algo3 20 34 -1.04756 0.08194
#>   Inst.28 algo2 algo4 20 37 -0.22502 0.07811
#>   Inst.28 algo3 algo4 34 37  0.82254 0.09980
#>   Inst.80 algo1 algo2 20 20  0.41685 0.05859
#>   Inst.80 algo1 algo3 20 31 -0.90375 0.07246
#>   Inst.80 algo1 algo4 20 47  0.08600 0.08118
#>   Inst.80 algo2 algo3 20 31 -1.32060 0.08540
#>   Inst.80 algo2 algo4 20 47 -0.33085 0.08937
#>   Inst.80 algo3 algo4 31 47  0.98975 0.09991
#>   Inst.22 algo1 algo2 20 20  0.42523 0.05721
#>   Inst.22 algo1 algo3 20 31 -0.89414 0.07134
#>   Inst.22 algo1 algo4 20 47  0.08368 0.08126
#>   Inst.22 algo2 algo3 20 31 -1.31938 0.08424
#>   Inst.22 algo2 algo4 20 47 -0.34155 0.08927
#>   Inst.22 algo3 algo4 31 47  0.97782 0.09978
#>    Inst.9 algo1 algo2 20 20  0.42781 0.05661
#>    Inst.9 algo1 algo3 20 35 -0.93116 0.07435
#>    Inst.9 algo1 algo4 20 44  0.11726 0.07795
#>    Inst.9 algo2 algo3 20 35 -1.35897 0.08684
#>    Inst.9 algo2 algo4 20 44 -0.31056 0.08606
#>    Inst.9 algo3 algo4 35 44  1.04842 0.09986
#>    Inst.5 algo1 algo2 20 20  0.42448 0.05732
#>    Inst.5 algo1 algo3 20 31 -0.89437 0.07139
#>    Inst.5 algo1 algo4 20 47  0.08329 0.08129
#>    Inst.5 algo2 algo3 20 31 -1.31885 0.08429
#>    Inst.5 algo2 algo4 20 47 -0.34119 0.08930
#>    Inst.5 algo3 algo4 31 47  0.97765 0.09978
#>   Inst.38 algo1 algo2 20 20  0.42535 0.05719
#>   Inst.38 algo1 algo3 20 31 -0.89411 0.07133
#>   Inst.38 algo1 algo4 20 47  0.08374 0.08126
#>   Inst.38 algo2 algo3 20 31 -1.31945 0.08423
#>   Inst.38 algo2 algo4 20 47 -0.34161 0.08926
#>   Inst.38 algo3 algo4 31 47  0.97785 0.09978
#>   Inst.16 algo1 algo2 20 20  0.42858 0.05651
#>   Inst.16 algo1 algo3 20 35 -0.93101 0.07431
#>   Inst.16 algo1 algo4 20 44  0.11772 0.07792
#>   Inst.16 algo2 algo3 20 35 -1.35959 0.08681
#>   Inst.16 algo2 algo4 20 44 -0.31086 0.08602
#>   Inst.16 algo3 algo4 35 44  1.04873 0.09986
#>    Inst.4 algo1 algo2 20 20  0.42983 0.05636
#>    Inst.4 algo1 algo3 20 35 -0.93076 0.07426
#>    Inst.4 algo1 algo4 20 44  0.11848 0.07786
#>    Inst.4 algo2 algo3 20 35 -1.36059 0.08675
#>    Inst.4 algo2 algo4 20 44 -0.31135 0.08596
#>    Inst.4 algo3 algo4 35 44  1.04924 0.09986
#>   Inst.86 algo1 algo2 20 20  0.42170 0.05776
#>   Inst.86 algo1 algo3 20 31 -0.89519 0.07161
#>   Inst.86 algo1 algo4 20 47  0.08184 0.08139
#>   Inst.86 algo2 algo3 20 31 -1.31689 0.08453
#>   Inst.86 algo2 algo4 20 47 -0.33986 0.08946
#>   Inst.86 algo3 algo4 31 47  0.97703 0.09981
#>   Inst.90 algo1 algo2 20 20  0.42351 0.05747
#>   Inst.90 algo1 algo3 20 31 -0.89465 0.07147
#>   Inst.90 algo1 algo4 20 47  0.08278 0.08133
#>   Inst.90 algo2 algo3 20 31 -1.31816 0.08437
#>   Inst.90 algo2 algo4 20 47 -0.34073 0.08936
#>   Inst.90 algo3 algo4 31 47  0.97743 0.09979
#>   Inst.70 algo1 algo2 20 20  0.42727 0.05668
#>   Inst.70 algo1 algo3 20 35 -0.93127 0.07438
#>   Inst.70 algo1 algo4 20 44  0.11693 0.07797
#>   Inst.70 algo2 algo3 20 35 -1.35854 0.08687
#>   Inst.70 algo2 algo4 20 44 -0.31034 0.08608
#>   Inst.70 algo3 algo4 35 44  1.04820 0.09986
#>   Inst.79 algo1 algo2 20 20  0.41771 0.05845
#>   Inst.79 algo1 algo3 20 31 -0.89637 0.07194
#>   Inst.79 algo1 algo4 20 47  0.07975 0.08156
#>   Inst.79 algo2 algo3 20 31 -1.31409 0.08491
#>   Inst.79 algo2 algo4 20 47 -0.33796 0.08972
#>   Inst.79 algo3 algo4 31 47  0.97613 0.09986
#>   Inst.78 algo1 algo2 20 20  0.42398 0.05740
#>   Inst.78 algo1 algo3 20 31 -0.89451 0.07143
#>   Inst.78 algo1 algo4 20 47  0.08303 0.08131
#>   Inst.78 algo2 algo3 20 31 -1.31850 0.08433
#>   Inst.78 algo2 algo4 20 47 -0.34095 0.08933
#>   Inst.78 algo3 algo4 31 47  0.97754 0.09979
#>   Inst.14 algo1 algo2 20 20  0.42735 0.05667
#>   Inst.14 algo1 algo3 20 35 -0.93125 0.07437
#>   Inst.14 algo1 algo4 20 44  0.11698 0.07797
#>   Inst.14 algo2 algo3 20 35 -1.35860 0.08686
#>   Inst.14 algo2 algo4 20 44 -0.31037 0.08608
#>   Inst.14 algo3 algo4 35 44  1.04823 0.09986
#>   Inst.56 algo1 algo2 20 20  0.42652 0.05702
#>   Inst.56 algo1 algo3 20 31 -0.89376 0.07124
#>   Inst.56 algo1 algo4 20 47  0.08435 0.08122
#>   Inst.56 algo2 algo3 20 31 -1.32028 0.08414
#>   Inst.56 algo2 algo4 20 47 -0.34217 0.08921
#>   Inst.56 algo3 algo4 31 47  0.97812 0.09977
#>   Inst.62 algo1 algo2 20 20  0.42367 0.05745
#>   Inst.62 algo1 algo3 20 31 -0.89461 0.07146
#>   Inst.62 algo1 algo4 20 47  0.08286 0.08132
#>   Inst.62 algo2 algo3 20 31 -1.31827 0.08436
#>   Inst.62 algo2 algo4 20 47 -0.34080 0.08935
#>   Inst.62 algo3 algo4 31 47  0.97747 0.09979
#>   Inst.93 algo1 algo2 20 20  0.23852 0.04651
#>   Inst.93 algo1 algo3 20 27 -0.75388 0.06930
#>   Inst.93 algo1 algo4 20 42  0.01106 0.08080
#>   Inst.93 algo2 algo3 20 27 -0.99240 0.07623
#>   Inst.93 algo2 algo4 20 42 -0.22745 0.08527
#>   Inst.93 algo3 algo4 27 42  0.76495 0.09963
#> 
#> #====================================