print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE, breakrows = FALSE)

Arguments

x

list object of class CAISEr (generated by run_experiment())

...

other parameters to be passed down to specific summary functions (currently unused)

echo

logical flag: should the print method actually print to screen?

digits

the minimum number of significant digits to be used. See print.default().

right

logical, indicating whether or not strings should be right-aligned.

breakrows

logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)
#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..
my.results
#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2     Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.2596 0.05493
#>   Inst.28 algo1 algo3 20 27 -0.8825 0.07431
#>   Inst.28 algo1 algo4 20 37 -0.2413 0.08051
#>   Inst.28 algo2 algo3 20 27 -1.1422 0.08453
#>   Inst.28 algo2 algo4 20 37 -0.5009 0.08856
#>   Inst.28 algo3 algo4 27 37  0.6413 0.09958
#>   Inst.80 algo1 algo2 20 20  0.2526 0.05237
#>   Inst.80 algo1 algo3 20 28 -0.9370 0.07395
#>   Inst.80 algo1 algo4 20 37 -0.2021 0.08099
#>   Inst.80 algo2 algo3 20 28 -1.1896 0.08190
#>   Inst.80 algo2 algo4 20 37 -0.4546 0.08670
#>   Inst.80 algo3 algo4 28 37  0.7349 0.09939
#>   Inst.22 algo1 algo2 20 20  0.2568 0.05388
#>   Inst.22 algo1 algo3 20 28 -0.9323 0.07454
#>   Inst.22 algo1 algo4 20 37 -0.2015 0.08060
#>   Inst.22 algo2 algo3 20 28 -1.1891 0.08397
#>   Inst.22 algo2 algo4 20 37 -0.4582 0.08774
#>   Inst.22 algo3 algo4 28 37  0.7308 0.09990
#>    Inst.9 algo1 algo2 20 20  0.2592 0.05472
#>    Inst.9 algo1 algo3 20 27 -0.8827 0.07424
#>    Inst.9 algo1 algo4 20 37 -0.2416 0.08052
#>    Inst.9 algo2 algo3 20 27 -1.1419 0.08428
#>    Inst.9 algo2 algo4 20 37 -0.5008 0.08840
#>    Inst.9 algo3 algo4 27 37  0.6411 0.09950
#>    Inst.5 algo1 algo2 20 20  0.2564 0.05373
#>    Inst.5 algo1 algo3 20 28 -0.9324 0.07448
#>    Inst.5 algo1 algo4 20 37 -0.2018 0.08062
#>    Inst.5 algo2 algo3 20 28 -1.1888 0.08378
#>    Inst.5 algo2 algo4 20 37 -0.4582 0.08763
#>    Inst.5 algo3 algo4 28 37  0.7307 0.09985
#>   Inst.38 algo1 algo2 20 20  0.2568 0.05390
#>   Inst.38 algo1 algo3 20 28 -0.9323 0.07455
#>   Inst.38 algo1 algo4 20 37 -0.2014 0.08059
#>   Inst.38 algo2 algo3 20 28 -1.1891 0.08400
#>   Inst.38 algo2 algo4 20 37 -0.4582 0.08776
#>   Inst.38 algo3 algo4 28 37  0.7308 0.09991
#>   Inst.16 algo1 algo2 20 20  0.2595 0.05489
#>   Inst.16 algo1 algo3 20 27 -0.8826 0.07430
#>   Inst.16 algo1 algo4 20 37 -0.2413 0.08051
#>   Inst.16 algo2 algo3 20 27 -1.1421 0.08448
#>   Inst.16 algo2 algo4 20 37 -0.5009 0.08853
#>   Inst.16 algo3 algo4 27 37  0.6412 0.09956
#>    Inst.4 algo1 algo2 20 20  0.2593 0.05497
#>    Inst.4 algo1 algo3 20 29 -0.9122 0.07483
#>    Inst.4 algo1 algo4 20 36 -0.2205 0.07964
#>    Inst.4 algo2 algo3 20 29 -1.1716 0.08518
#>    Inst.4 algo2 algo4 20 36 -0.4798 0.08786
#>    Inst.4 algo3 algo4 29 36  0.6917 0.09947
#>   Inst.86 algo1 algo2 20 20  0.2550 0.05320
#>   Inst.86 algo1 algo3 20 28 -0.9374 0.07435
#>   Inst.86 algo1 algo4 20 37 -0.1988 0.08077
#>   Inst.86 algo2 algo3 20 28 -1.1923 0.08314
#>   Inst.86 algo2 algo4 20 37 -0.4538 0.08729
#>   Inst.86 algo3 algo4 28 37  0.7386 0.09976
#>   Inst.90 algo1 algo2 20 20  0.2559 0.05354
#>   Inst.90 algo1 algo3 20 28 -0.9326 0.07441
#>   Inst.90 algo1 algo4 20 37 -0.2022 0.08065
#>   Inst.90 algo2 algo3 20 28 -1.1885 0.08353
#>   Inst.90 algo2 algo4 20 37 -0.4581 0.08749
#>   Inst.90 algo3 algo4 28 37  0.7305 0.09977
#>   Inst.70 algo1 algo2 20 20  0.2589 0.05460
#>   Inst.70 algo1 algo3 20 27 -0.8829 0.07420
#>   Inst.70 algo1 algo4 20 37 -0.2419 0.08053
#>   Inst.70 algo2 algo3 20 27 -1.1417 0.08414
#>   Inst.70 algo2 algo4 20 37 -0.5007 0.08831
#>   Inst.70 algo3 algo4 27 37  0.6410 0.09945
#>   Inst.79 algo1 algo2 20 20  0.2530 0.05250
#>   Inst.79 algo1 algo3 20 28 -0.9386 0.07410
#>   Inst.79 algo1 algo4 20 37 -0.2000 0.08092
#>   Inst.79 algo2 algo3 20 28 -1.1916 0.08219
#>   Inst.79 algo2 algo4 20 37 -0.4530 0.08676
#>   Inst.79 algo3 algo4 28 37  0.7386 0.09949
#>   Inst.78 algo1 algo2 20 20  0.2562 0.05363
#>   Inst.78 algo1 algo3 20 28 -0.9325 0.07445
#>   Inst.78 algo1 algo4 20 37 -0.2020 0.08063
#>   Inst.78 algo2 algo3 20 28 -1.1887 0.08365
#>   Inst.78 algo2 algo4 20 37 -0.4581 0.08756
#>   Inst.78 algo3 algo4 28 37  0.7306 0.09981
#>   Inst.14 algo1 algo2 20 20  0.2589 0.05462
#>   Inst.14 algo1 algo3 20 27 -0.8828 0.07421
#>   Inst.14 algo1 algo4 20 37 -0.2418 0.08053
#>   Inst.14 algo2 algo3 20 27 -1.1418 0.08416
#>   Inst.14 algo2 algo4 20 37 -0.5008 0.08832
#>   Inst.14 algo3 algo4 27 37  0.6410 0.09946
#>   Inst.56 algo1 algo2 20 20  0.2581 0.05427
#>   Inst.56 algo1 algo3 20 28 -0.9343 0.07479
#>   Inst.56 algo1 algo4 20 38 -0.1914 0.07953
#>   Inst.56 algo2 algo3 20 28 -1.1924 0.08445
#>   Inst.56 algo2 algo4 20 38 -0.4495 0.08699
#>   Inst.56 algo3 algo4 28 38  0.7429 0.09931
#>   Inst.62 algo1 algo2 20 20  0.2560 0.05357
#>   Inst.62 algo1 algo3 20 28 -0.9326 0.07442
#>   Inst.62 algo1 algo4 20 37 -0.2021 0.08064
#>   Inst.62 algo2 algo3 20 28 -1.1886 0.08357
#>   Inst.62 algo2 algo4 20 37 -0.4581 0.08752
#>   Inst.62 algo3 algo4 28 37  0.7305 0.09978
#>   Inst.93 algo1 algo2 20 20  0.2599 0.05506
#>   Inst.93 algo1 algo3 20 27 -0.8824 0.07436
#>   Inst.93 algo1 algo4 20 37 -0.2410 0.08051
#>   Inst.93 algo2 algo3 20 27 -1.1423 0.08469
#>   Inst.93 algo2 algo4 20 37 -0.5010 0.08866
#>   Inst.93 algo3 algo4 27 37  0.6414 0.09963
#> 
#> #====================================