print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : .
#> Sampling algorithms on instance Inst.79 : .
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : .
#> Sampling algorithms on instance Inst.93 : .

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.33709 0.04292
#>   Inst.28 algo1 algo3 20 33 -0.87840 0.07180
#>   Inst.28 algo1 algo4 20 39  0.02114 0.07311
#>   Inst.28 algo2 algo3 20 33 -1.21549 0.08258
#>   Inst.28 algo2 algo4 20 39 -0.31596 0.08116
#>   Inst.28 algo3 algo4 33 39  0.89953 0.09968
#>   Inst.80 algo1 algo2 20 20  0.34567 0.04335
#>   Inst.80 algo1 algo3 20 34 -0.85674 0.07090
#>   Inst.80 algo1 algo4 20 42  0.01131 0.07435
#>   Inst.80 algo2 algo3 20 34 -1.20241 0.08211
#>   Inst.80 algo2 algo4 20 42 -0.33436 0.08259
#>   Inst.80 algo3 algo4 34 42  0.86804 0.09994
#>   Inst.22 algo1 algo2 20 20  0.33930 0.04291
#>   Inst.22 algo1 algo3 20 33 -0.90075 0.07155
#>   Inst.22 algo1 algo4 20 39  0.04073 0.07289
#>   Inst.22 algo2 algo3 20 33 -1.24004 0.08249
#>   Inst.22 algo2 algo4 20 39 -0.29856 0.08097
#>   Inst.22 algo3 algo4 33 39  0.94148 0.09952
#>    Inst.9 algo1 algo2 20 20  0.33753 0.04291
#>    Inst.9 algo1 algo3 20 33 -0.87890 0.07180
#>    Inst.9 algo1 algo4 20 39  0.02061 0.07309
#>    Inst.9 algo2 algo3 20 33 -1.21644 0.08260
#>    Inst.9 algo2 algo4 20 39 -0.31693 0.08116
#>    Inst.9 algo3 algo4 33 39  0.89951 0.09968
#>    Inst.5 algo1 algo2 20 20  0.33964 0.04291
#>    Inst.5 algo1 algo3 20 33 -0.90114 0.07156
#>    Inst.5 algo1 algo4 20 39  0.04032 0.07288
#>    Inst.5 algo2 algo3 20 33 -1.24078 0.08251
#>    Inst.5 algo2 algo4 20 39 -0.29932 0.08097
#>    Inst.5 algo3 algo4 33 39  0.94147 0.09953
#>   Inst.38 algo1 algo2 20 20  0.25065 0.05996
#>   Inst.38 algo1 algo3 20 26 -0.82122 0.07106
#>   Inst.38 algo1 algo4 20 39 -0.05460 0.07814
#>   Inst.38 algo2 algo3 20 26 -1.07186 0.08721
#>   Inst.38 algo2 algo4 20 39 -0.30525 0.09141
#>   Inst.38 algo3 algo4 26 39  0.76662 0.09857
#>   Inst.16 algo1 algo2 20 20  0.25116 0.05964
#>   Inst.16 algo1 algo3 20 26 -0.81917 0.07125
#>   Inst.16 algo1 algo4 20 39 -0.05353 0.07806
#>   Inst.16 algo2 algo3 20 26 -1.07034 0.08720
#>   Inst.16 algo2 algo4 20 39 -0.30470 0.09118
#>   Inst.16 algo3 algo4 26 39  0.76564 0.09870
#>    Inst.4 algo1 algo2 20 20  0.25132 0.05954
#>    Inst.4 algo1 algo3 20 26 -0.81854 0.07131
#>    Inst.4 algo1 algo4 20 39 -0.05320 0.07803
#>    Inst.4 algo2 algo3 20 26 -1.06986 0.08721
#>    Inst.4 algo2 algo4 20 39 -0.30453 0.09111
#>    Inst.4 algo3 algo4 26 39  0.76534 0.09875
#>   Inst.86 algo1 algo2 20 20  0.25005 0.06023
#>   Inst.86 algo1 algo3 20 26 -0.84295 0.07159
#>   Inst.86 algo1 algo4 20 39 -0.03735 0.07827
#>   Inst.86 algo2 algo3 20 26 -1.09300 0.08783
#>   Inst.86 algo2 algo4 20 39 -0.28740 0.09160
#>   Inst.86 algo3 algo4 26 39  0.80560 0.09911
#>   Inst.90 algo1 algo2 20 20  0.25074 0.06020
#>   Inst.90 algo1 algo3 20 26 -0.82047 0.07082
#>   Inst.90 algo1 algo4 20 38 -0.05264 0.08012
#>   Inst.90 algo2 algo3 20 26 -1.07121 0.08717
#>   Inst.90 algo2 algo4 20 38 -0.30339 0.09321
#>   Inst.90 algo3 algo4 26 38  0.76783 0.09994
#>   Inst.70 algo1 algo2 20 20  0.28486 0.06097
#>   Inst.70 algo1 algo3 20 20 -0.90941 0.06408
#>   Inst.70 algo1 algo4 20 27  0.11885 0.08229
#>   Inst.70 algo2 algo3 20 20 -1.19427 0.08324
#>   Inst.70 algo2 algo4 20 27 -0.16601 0.09557
#>   Inst.70 algo3 algo4 20 27  1.02826 0.09872
#>   Inst.79 algo1 algo2 20 20  0.29201 0.06171
#>   Inst.79 algo1 algo3 20 20 -0.90284 0.06459
#>   Inst.79 algo1 algo4 20 27  0.12768 0.08254
#>   Inst.79 algo2 algo3 20 20 -1.19485 0.08370
#>   Inst.79 algo2 algo4 20 27 -0.16433 0.09576
#>   Inst.79 algo3 algo4 20 27  1.03052 0.09886
#>   Inst.78 algo1 algo2 20 20  0.28769 0.06121
#>   Inst.78 algo1 algo3 20 20 -0.90680 0.06422
#>   Inst.78 algo1 algo4 20 27  0.12235 0.08236
#>   Inst.78 algo2 algo3 20 20 -1.19450 0.08337
#>   Inst.78 algo2 algo4 20 27 -0.16534 0.09562
#>   Inst.78 algo3 algo4 20 27  1.02916 0.09875
#>   Inst.14 algo1 algo2 20 20  0.28480 0.06097
#>   Inst.14 algo1 algo3 20 20 -0.90946 0.06408
#>   Inst.14 algo1 algo4 20 27  0.11878 0.08229
#>   Inst.14 algo2 algo3 20 20 -1.19426 0.08324
#>   Inst.14 algo2 algo4 20 27 -0.16602 0.09557
#>   Inst.14 algo3 algo4 20 27  1.02824 0.09872
#>   Inst.56 algo1 algo2 20 20  0.28595 0.06105
#>   Inst.56 algo1 algo3 20 20 -0.90841 0.06413
#>   Inst.56 algo1 algo4 20 27  0.12020 0.08231
#>   Inst.56 algo2 algo3 20 20 -1.19436 0.08329
#>   Inst.56 algo2 algo4 20 27 -0.16575 0.09559
#>   Inst.56 algo3 algo4 20 27  1.02861 0.09873
#>   Inst.62 algo1 algo2 20 20  0.28791 0.06123
#>   Inst.62 algo1 algo3 20 20 -0.90660 0.06424
#>   Inst.62 algo1 algo4 20 27  0.12262 0.08236
#>   Inst.62 algo2 algo3 20 20 -1.19452 0.08339
#>   Inst.62 algo2 algo4 20 27 -0.16529 0.09563
#>   Inst.62 algo3 algo4 20 27  1.02923 0.09875
#>   Inst.93 algo1 algo2 20 20  0.28340 0.06088
#>   Inst.93 algo1 algo3 20 20 -0.91075 0.06404
#>   Inst.93 algo1 algo4 20 27  0.11706 0.08227
#>   Inst.93 algo2 algo3 20 20 -1.19415 0.08320
#>   Inst.93 algo2 algo4 20 27 -0.16635 0.09556
#>   Inst.93 algo3 algo4 20 27  1.02780 0.09873
#> 
#> #====================================