print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 6, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object of class xtable containint a LaTeX-formatted summary

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ...
#> Sampling algorithms on instance Inst.80 : .
#> Sampling algorithms on instance Inst.22 : .
#> Sampling algorithms on instance Inst.9 : .
#> Sampling algorithms on instance Inst.5 : ...
#> Sampling algorithms on instance Inst.38 : .
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : .
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ...
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ...
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ...

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2         Phi        SE
#>   Inst.28 algo1 algo2 20 20  0.30135611 0.0513877
#>   Inst.28 algo1 algo3 20 41 -0.89607296 0.0743116
#>   Inst.28 algo1 algo4 20 40  0.03763992 0.0696286
#>   Inst.28 algo2 algo3 20 41 -1.19742908 0.0895029
#>   Inst.28 algo2 algo4 20 40 -0.26371619 0.0831964
#>   Inst.28 algo3 algo4 41 40  0.93371288 0.0993461
#>   Inst.80 algo1 algo2 20 20  0.22022455 0.0672364
#>   Inst.80 algo1 algo3 20 22 -0.82231502 0.0678596
#>   Inst.80 algo1 algo4 20 34 -0.00473195 0.0772926
#>   Inst.80 algo2 algo3 20 22 -1.04253957 0.0929724
#>   Inst.80 algo2 algo4 20 34 -0.22495650 0.0985140
#>   Inst.80 algo3 algo4 22 34  0.81758307 0.0988977
#>   Inst.22 algo1 algo2 20 20  0.37184808 0.0563262
#>   Inst.22 algo1 algo3 20 28 -0.71688951 0.0776416
#>   Inst.22 algo1 algo4 20 29 -0.07926681 0.0769530
#>   Inst.22 algo2 algo3 20 28 -1.08873759 0.0870713
#>   Inst.22 algo2 algo4 20 29 -0.45111489 0.0842763
#>   Inst.22 algo3 algo4 28 29  0.63762270 0.0991086
#>    Inst.9 algo1 algo2 20 20  0.29457130 0.0640940
#>    Inst.9 algo1 algo3 20 23 -0.81286230 0.0693580
#>    Inst.9 algo1 algo4 20 36  0.00840838 0.0818620
#>    Inst.9 algo2 algo3 20 23 -1.10743361 0.0879882
#>    Inst.9 algo2 algo4 20 36 -0.28616292 0.0960579
#>    Inst.9 algo3 algo4 23 36  0.82127068 0.0997257
#>    Inst.5 algo1 algo2 20 20  0.26278505 0.0606019
#>    Inst.5 algo1 algo3 20 37 -0.76630171 0.0732465
#>    Inst.5 algo1 algo4 20 44 -0.06564464 0.0770274
#>    Inst.5 algo2 algo3 20 37 -1.02908677 0.0898620
#>    Inst.5 algo2 algo4 20 44 -0.32842969 0.0913293
#>    Inst.5 algo3 algo4 37 44  0.70065707 0.0996121
#>   Inst.38 algo1 algo2 20 20  0.29963863 0.0589293
#>   Inst.38 algo1 algo3 20 23 -0.85759646 0.0677700
#>   Inst.38 algo1 algo4 20 31  0.06248101 0.0758253
#>   Inst.38 algo2 algo3 20 23 -1.15723509 0.0882698
#>   Inst.38 algo2 algo4 20 31 -0.23715762 0.0922917
#>   Inst.38 algo3 algo4 23 31  0.92007747 0.0987461
#>   Inst.16 algo1 algo2 20 20  0.25586059 0.0460065
#>   Inst.16 algo1 algo3 20 29 -0.86412110 0.0709336
#>   Inst.16 algo1 algo4 20 40 -0.07522312 0.0763743
#>   Inst.16 algo2 algo3 20 29 -1.11998169 0.0805888
#>   Inst.16 algo2 algo4 20 40 -0.33108371 0.0837119
#>   Inst.16 algo3 algo4 29 40  0.78889798 0.0990067
#>    Inst.4 algo1 algo2 20 20  0.24332668 0.0534080
#>    Inst.4 algo1 algo3 20 22 -1.00835358 0.0758112
#>    Inst.4 algo1 algo4 20 22 -0.03291983 0.0722940
#>    Inst.4 algo2 algo3 20 22 -1.25168026 0.0888005
#>    Inst.4 algo2 algo4 20 22 -0.27624650 0.0837276
#>    Inst.4 algo3 algo4 22 22  0.97543376 0.0992156
#>   Inst.86 algo1 algo2 20 24  0.36201209 0.0673946
#>   Inst.86 algo1 algo3 20 24 -0.82259606 0.0682737
#>   Inst.86 algo1 algo4 20 38 -0.03802671 0.0767129
#>   Inst.86 algo2 algo3 24 24 -1.18460815 0.0949341
#>   Inst.86 algo2 algo4 24 38 -0.40003880 0.0987413
#>   Inst.86 algo3 algo4 24 38  0.78456935 0.0989546
#>   Inst.90 algo1 algo2 20 20  0.27542936 0.0519020
#>   Inst.90 algo1 algo3 20 34 -0.85969748 0.0735961
#>   Inst.90 algo1 algo4 20 47  0.09491224 0.0782572
#>   Inst.90 algo2 algo3 20 34 -1.13512683 0.0813135
#>   Inst.90 algo2 algo4 20 47 -0.18051712 0.0832617
#>   Inst.90 algo3 algo4 34 47  0.95460971 0.0990643
#>   Inst.70 algo1 algo2 20 20  0.29911612 0.0646918
#>   Inst.70 algo1 algo3 20 33 -0.92619261 0.0797443
#>   Inst.70 algo1 algo4 20 34 -0.00138446 0.0741139
#>   Inst.70 algo2 algo3 20 33 -1.22530873 0.0956667
#>   Inst.70 algo2 algo4 20 34 -0.30050058 0.0884471
#>   Inst.70 algo3 algo4 33 34  0.92480815 0.0999709
#>   Inst.79 algo1 algo2 20 20  0.30208626 0.0415983
#>   Inst.79 algo1 algo3 20 37 -0.82892240 0.0716237
#>   Inst.79 algo1 algo4 20 45  0.06696660 0.0751496
#>   Inst.79 algo2 algo3 20 37 -1.13100865 0.0784963
#>   Inst.79 algo2 algo4 20 45 -0.23511965 0.0794315
#>   Inst.79 algo3 algo4 37 45  0.89588900 0.0990575
#>   Inst.78 algo1 algo2 20 20  0.36118165 0.0521165
#>   Inst.78 algo1 algo3 20 20 -0.82586338 0.0612370
#>   Inst.78 algo1 algo4 20 21 -0.05051365 0.0864034
#>   Inst.78 algo2 algo3 20 20 -1.18704503 0.0749130
#>   Inst.78 algo2 algo4 20 21 -0.41169530 0.0943762
#>   Inst.78 algo3 algo4 20 21  0.77534973 0.0992504
#>   Inst.14 algo1 algo2 20 20  0.32700900 0.0595986
#>   Inst.14 algo1 algo3 20 21 -0.96836082 0.0721332
#>   Inst.14 algo1 algo4 20 27  0.02831483 0.0750828
#>   Inst.14 algo2 algo3 20 21 -1.29536982 0.0899163
#>   Inst.14 algo2 algo4 20 27 -0.29869417 0.0895164
#>   Inst.14 algo3 algo4 21 27  0.99667565 0.0985968
#>   Inst.56 algo1 algo2 20 20  0.21832737 0.0560057
#>   Inst.56 algo1 algo3 20 25 -0.92284743 0.0721440
#>   Inst.56 algo1 algo4 20 32  0.04784007 0.0749301
#>   Inst.56 algo2 algo3 20 25 -1.14117480 0.0867267
#>   Inst.56 algo2 algo4 20 32 -0.17048730 0.0872327
#>   Inst.56 algo3 algo4 25 32  0.97068750 0.0987955
#>   Inst.62 algo1 algo2 20 20  0.24211158 0.0490293
#>   Inst.62 algo1 algo3 20 28 -0.81855401 0.0683251
#>   Inst.62 algo1 algo4 20 45 -0.04459220 0.0786032
#>   Inst.62 algo2 algo3 20 28 -1.06066559 0.0792886
#>   Inst.62 algo2 algo4 20 45 -0.28670378 0.0867414
#>   Inst.62 algo3 algo4 28 45  0.77396181 0.0985884
#>   Inst.93 algo1 algo2 20 20  0.34233999 0.0671274
#>   Inst.93 algo1 algo3 20 35 -0.84457216 0.0725656
#>   Inst.93 algo1 algo4 20 51  0.01479201 0.0782264
#>   Inst.93 algo2 algo3 20 35 -1.18691214 0.0937123
#>   Inst.93 algo2 algo4 20 51 -0.32754797 0.0955614
#>   Inst.93 algo3 algo4 35 51  0.85936417 0.0996077
#> 
#> #====================================