print.CAISEr

S3 method for printing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
print(x, ..., echo = TRUE, digits = 4, right = TRUE,
  breakrows = FALSE)

Arguments

x
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

echo
logical flag: should the print method actually print to screen?

digits
the minimum number of significant digits to be used. See print.default().

right
logical, indicating whether or not strings should be right-aligned.

breakrows
logical, indicating whether to "widen" the output table by placing the bottom half to the right of the top half.

Value

data frame object containing the summary table (invisibly)

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : .
#> Sampling algorithms on instance Inst.80 : .
#> Sampling algorithms on instance Inst.22 : .
#> Sampling algorithms on instance Inst.9 : .
#> Sampling algorithms on instance Inst.5 : .
#> Sampling algorithms on instance Inst.38 : .
#> Sampling algorithms on instance Inst.16 : .
#> Sampling algorithms on instance Inst.4 : .
#> Sampling algorithms on instance Inst.86 : .
#> Sampling algorithms on instance Inst.90 : .
#> Sampling algorithms on instance Inst.70 : .
#> Sampling algorithms on instance Inst.79 : .
#> Sampling algorithms on instance Inst.78 : .
#> Sampling algorithms on instance Inst.14 : .
#> Sampling algorithms on instance Inst.56 : .
#> Sampling algorithms on instance Inst.62 : .
#> Sampling algorithms on instance Inst.93 : .

my.results

#> #====================================
#>  Summary table of CAISEr object
#>  Instance  Alg1  Alg2 N1 N2      Phi      SE
#>   Inst.28 algo1 algo2 20 20  0.34163 0.05851
#>   Inst.28 algo1 algo3 20 25 -0.74497 0.07669
#>   Inst.28 algo1 algo4 20 28  0.04279 0.07393
#>   Inst.28 algo2 algo3 20 25 -1.08660 0.09085
#>   Inst.28 algo2 algo4 20 28 -0.29884 0.08608
#>   Inst.28 algo3 algo4 25 28  0.78776 0.09970
#>   Inst.80 algo1 algo2 20 20  0.33878 0.05762
#>   Inst.80 algo1 algo3 20 25 -0.78176 0.07144
#>   Inst.80 algo1 algo4 20 33  0.07555 0.07786
#>   Inst.80 algo2 algo3 20 25 -1.12053 0.08588
#>   Inst.80 algo2 algo4 20 33 -0.26322 0.08878
#>   Inst.80 algo3 algo4 25 33  0.85731 0.09898
#>   Inst.22 algo1 algo2 20 20  0.34108 0.05818
#>   Inst.22 algo1 algo3 20 25 -0.78606 0.07161
#>   Inst.22 algo1 algo4 20 32  0.08671 0.07876
#>   Inst.22 algo2 algo3 20 25 -1.12714 0.08654
#>   Inst.22 algo2 algo4 20 32 -0.25436 0.08996
#>   Inst.22 algo3 algo4 25 32  0.87277 0.09996
#>    Inst.9 algo1 algo2 20 20  0.34149 0.05843
#>    Inst.9 algo1 algo3 20 25 -0.74448 0.07664
#>    Inst.9 algo1 algo4 20 28  0.04279 0.07396
#>    Inst.9 algo2 algo3 20 25 -1.08597 0.09078
#>    Inst.9 algo2 algo4 20 28 -0.29870 0.08607
#>    Inst.9 algo3 algo4 25 28  0.78727 0.09971
#>    Inst.5 algo1 algo2 20 20  0.34098 0.05814
#>    Inst.5 algo1 algo3 20 25 -0.78592 0.07161
#>    Inst.5 algo1 algo4 20 32  0.08700 0.07878
#>    Inst.5 algo2 algo3 20 25 -1.12689 0.08651
#>    Inst.5 algo2 algo4 20 32 -0.25398 0.08995
#>    Inst.5 algo3 algo4 25 32  0.87291 0.09997
#>   Inst.38 algo1 algo2 20 20  0.34109 0.05819
#>   Inst.38 algo1 algo3 20 25 -0.78608 0.07161
#>   Inst.38 algo1 algo4 20 32  0.08667 0.07876
#>   Inst.38 algo2 algo3 20 25 -1.12717 0.08655
#>   Inst.38 algo2 algo4 20 32 -0.25442 0.08997
#>   Inst.38 algo3 algo4 25 32  0.87275 0.09995
#>   Inst.16 algo1 algo2 20 20  0.34160 0.05850
#>   Inst.16 algo1 algo3 20 25 -0.74487 0.07668
#>   Inst.16 algo1 algo4 20 28  0.04279 0.07393
#>   Inst.16 algo2 algo3 20 25 -1.08647 0.09084
#>   Inst.16 algo2 algo4 20 28 -0.29881 0.08607
#>   Inst.16 algo3 algo4 25 28  0.78766 0.09970
#>    Inst.4 algo1 algo2 20 20  0.34144 0.05855
#>    Inst.4 algo1 algo3 20 26 -0.74763 0.07434
#>    Inst.4 algo1 algo4 20 27  0.04161 0.07623
#>    Inst.4 algo2 algo3 20 26 -1.08907 0.08890
#>    Inst.4 algo2 algo4 20 27 -0.29982 0.08808
#>    Inst.4 algo3 algo4 26 27  0.78925 0.09964
#>   Inst.86 algo1 algo2 20 20  0.33941 0.05777
#>   Inst.86 algo1 algo3 20 25 -0.78263 0.07132
#>   Inst.86 algo1 algo4 20 33  0.07366 0.07771
#>   Inst.86 algo2 algo3 20 25 -1.12204 0.08600
#>   Inst.86 algo2 algo4 20 33 -0.26574 0.08886
#>   Inst.86 algo3 algo4 25 33  0.85630 0.09886
#>   Inst.90 algo1 algo2 20 20  0.34085 0.05808
#>   Inst.90 algo1 algo3 20 25 -0.78573 0.07161
#>   Inst.90 algo1 algo4 20 32  0.08736 0.07880
#>   Inst.90 algo2 algo3 20 25 -1.12658 0.08647
#>   Inst.90 algo2 algo4 20 32 -0.25348 0.08992
#>   Inst.90 algo3 algo4 25 32  0.87310 0.09999
#>   Inst.70 algo1 algo2 20 20  0.34142 0.05839
#>   Inst.70 algo1 algo3 20 25 -0.74420 0.07662
#>   Inst.70 algo1 algo4 20 28  0.04279 0.07398
#>   Inst.70 algo2 algo3 20 25 -1.08562 0.09074
#>   Inst.70 algo2 algo4 20 28 -0.29862 0.08606
#>   Inst.70 algo3 algo4 25 28  0.78699 0.09971
#>   Inst.79 algo1 algo2 20 20  0.30385 0.05726
#>   Inst.79 algo1 algo3 20 20 -0.88326 0.06906
#>   Inst.79 algo1 algo4 20 24 -0.03371 0.08114
#>   Inst.79 algo2 algo3 20 20 -1.18711 0.08289
#>   Inst.79 algo2 algo4 20 24 -0.33756 0.09103
#>   Inst.79 algo3 algo4 20 24  0.84955 0.09856
#>   Inst.78 algo1 algo2 20 20  0.30387 0.05714
#>   Inst.78 algo1 algo3 20 20 -0.88130 0.06972
#>   Inst.78 algo1 algo4 20 24 -0.03587 0.08060
#>   Inst.78 algo2 algo3 20 20 -1.18518 0.08377
#>   Inst.78 algo2 algo4 20 24 -0.33975 0.09086
#>   Inst.78 algo3 algo4 20 24  0.84543 0.09892
#>   Inst.14 algo1 algo2 20 20  0.30389 0.05713
#>   Inst.14 algo1 algo3 20 20 -0.87999 0.07023
#>   Inst.14 algo1 algo4 20 24 -0.03732 0.08028
#>   Inst.14 algo2 algo3 20 20 -1.18388 0.08442
#>   Inst.14 algo2 algo4 20 24 -0.34121 0.09077
#>   Inst.14 algo3 algo4 20 24  0.84267 0.09919
#>   Inst.56 algo1 algo2 20 20  0.30389 0.05713
#>   Inst.56 algo1 algo3 20 20 -0.88051 0.07002
#>   Inst.56 algo1 algo4 20 24 -0.03675 0.08040
#>   Inst.56 algo2 algo3 20 20 -1.18439 0.08416
#>   Inst.56 algo2 algo4 20 24 -0.34063 0.09080
#>   Inst.56 algo3 algo4 20 24  0.84376 0.09908
#>   Inst.62 algo1 algo2 20 20  0.30387 0.05714
#>   Inst.62 algo1 algo3 20 20 -0.88140 0.06969
#>   Inst.62 algo1 algo4 20 24 -0.03576 0.08063
#>   Inst.62 algo2 algo3 20 20 -1.18527 0.08373
#>   Inst.62 algo2 algo4 20 24 -0.33964 0.09086
#>   Inst.62 algo3 algo4 20 24  0.84564 0.09890
#>   Inst.93 algo1 algo2 20 20  0.30390 0.05714
#>   Inst.93 algo1 algo3 20 20 -0.87936 0.07049
#>   Inst.93 algo1 algo4 20 24 -0.03802 0.08013
#>   Inst.93 algo2 algo3 20 20 -1.18326 0.08474
#>   Inst.93 algo2 algo4 20 24 -0.34192 0.09074
#>   Inst.93 algo3 algo4 20 24  0.84134 0.09933
#> 
#> #====================================