summary.CAISEr

S3 method for summarizing CAISEr objects (the output of run_experiment()).

# S3 method for CAISEr
summary(object, ...)

Arguments

object
list object of class CAISEr (generated by run_experiment())

...
other parameters to be passed down to specific summary functions (currently unused)

Examples

# Example using dummy algorithms and instances. See ?dummyalgo for details.
# In this case all instances are the same, so we expect all cases to return
# a percent difference of approx. phi.j = 1.0 and sample sizes of
# approx. n1 = 31, n2 = 87
algorithm1 <- list(FUN = "dummyalgo", alias = "algo1",
                   distribution.fun = "rnorm",
                   distribution.pars = list(mean = 10, sd = 1))
algorithm2 <- list(FUN = "dummyalgo", alias = "algo2",
                   distribution.fun = "rnorm",
                   distribution.pars = list(mean = 20, sd = 4))
algolist <- list(algorithm1, algorithm2)
instlist <- vector(100, mode = "list")
for (i in 1:100) instlist[[i]] <- list(FUN = "dummyinstance",
                                       alias = paste0("Inst. ", i))

out <- run_experiment(Instance.list = instlist,
                      Algorithm.list = algolist,
                      power = 0.8,
                      d = 1,
                      sig.level = 0.01,
                      se.max = 0.05,
                      dif = "perc",
                      nmax   = 200,
                      seed   = 1234)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 16
#> Available number of instances: 100
#> -----------------------------
#> Sampling algorithms on instance: Inst. 12...
#> Sampling algorithms on instance: Inst. 62.....
#> Sampling algorithms on instance: Inst. 60.....
#> Sampling algorithms on instance: Inst. 61.....
#> Sampling algorithms on instance: Inst. 83....
#> Sampling algorithms on instance: Inst. 97....
#> Sampling algorithms on instance: Inst. 1....
#> Sampling algorithms on instance: Inst. 22.....
#> Sampling algorithms on instance: Inst. 99...
#> Sampling algorithms on instance: Inst. 47...
#> Sampling algorithms on instance: Inst. 63....
#> Sampling algorithms on instance: Inst. 49....
#> Sampling algorithms on instance: Inst. 25...
#> Sampling algorithms on instance: Inst. 81....
#> Sampling algorithms on instance: Inst. 26.....
#> Sampling algorithms on instance: Inst. 72.....

summary(out)

#> #====================================
#>  CAISEr object:
#>  Number of instances sampled: 16
#>  Number of instances required: 16
#>  Adequate power: TRUE
#>  Total number of runs of algo1: 486
#>  Total number of runs of algo2: 1324
#> #====================================