summary.CAISEr

S3 method for summarizing CAISEr objects output by run_experiment()). Input parameters test, alternative and sig.level can be used to override the ones used in the call to run_experiment().

# S3 method for CAISEr
summary(object, test = NULL, alternative = NULL,
  sig.level = NULL, ...)

Arguments

object
list object of class CAISEr (generated by run_experiment())

test
type of test to be used ("t.test", "wilcoxon" or "binomial")

alternative
type of alternative hypothesis ("two.sided" or "less" or "greater"). See calc_instances() for details.

sig.level
desired family-wise significance level (alpha) for the experiment

...
other parameters to be passed down to specific summary functions (currently unused)

Value

A list object is returned invisibly, containing the details of all tests performed as well as information on the total number of runs dedicated to each algorithm.

Examples

# Example using four dummy algorithms and 100 dummy instances.
# See [dummyalgo()] and [dummyinstance()] for details.
# Generating 4 dummy algorithms here, with means 15, 10, 30, 15 and standard
# deviations 2, 4, 6, 8.
algorithms <- mapply(FUN = function(i, m, s){
  list(FUN   = "dummyalgo",
       alias = paste0("algo", i),
       distribution.fun  = "rnorm",
       distribution.pars = list(mean = m, sd = s))},
  i = c(alg1 = 1, alg2 = 2, alg3 = 3, alg4 = 4),
  m = c(15, 10, 30, 15),
  s = c(2, 4, 6, 8),
  SIMPLIFY = FALSE)

# Generate 100 dummy instances with centered exponential distributions
instances <- lapply(1:100,
                    function(i) {rate <- runif(1, 1, 10)
                                 list(FUN   = "dummyinstance",
                                      alias = paste0("Inst.", i),
                                      distr = "rexp", rate = rate,
                                      bias  = -1 / rate)})

my.results <- run_experiment(instances, algorithms,
                             d = 1, se.max = .1,
                             power = .9, sig.level = .05,
                             power.target = "mean",
                             dif = "perc", comparisons = "all.vs.all",
                             seed = 1234, ncpus = 1)

#> CAISEr running
#> -----------------------------
#> Required number of instances: 17
#> Available number of instances: 100
#> Using 1 cores.
#> -----------------------------
#> Sampling algorithms on instance Inst.28 : ..
#> Sampling algorithms on instance Inst.80 : ..
#> Sampling algorithms on instance Inst.22 : ..
#> Sampling algorithms on instance Inst.9 : ..
#> Sampling algorithms on instance Inst.5 : ..
#> Sampling algorithms on instance Inst.38 : ..
#> Sampling algorithms on instance Inst.16 : ..
#> Sampling algorithms on instance Inst.4 : ..
#> Sampling algorithms on instance Inst.86 : ..
#> Sampling algorithms on instance Inst.90 : ..
#> Sampling algorithms on instance Inst.70 : ..
#> Sampling algorithms on instance Inst.79 : ..
#> Sampling algorithms on instance Inst.78 : ..
#> Sampling algorithms on instance Inst.14 : ..
#> Sampling algorithms on instance Inst.56 : ..
#> Sampling algorithms on instance Inst.62 : ..
#> Sampling algorithms on instance Inst.93 : ..

summary(my.results)

#> #====================================
#>  CAISEr object:
#>  Number of instances sampled: 17
#>  Number of instances required: 17
#>  Adequate power: TRUE
#>  Total runs of algo1 : 340
#>  Total runs of algo2 : 340
#>  Total runs of algo3 : 542
#>  Total runs of algo4 : 745
#> #====================================
#>  Pairwise comparisons of interest:
#>  Test: t.test
#>  H1: two.sided
#>  Comparisons: all.vs.all
#>  Alpha (FWER): 0.05
#>  Power target: mean
#>  Desired power: 0.9
#> #====================================
#> Tests using Holm's step-down procedure:
#> 
#>  Test 1 : algo1 x algo3
#>  H0: mean = 0
#>  alpha		= 0.008333 
#>  p-value	= 7.845e-29
#>  Est. mean	= -0.9429
#>  CI{1-alpha}	= [ -0.9567 -0.9292 ]
#>  d		= -50.02262
#> 
#>  Test 2 : algo2 x algo3
#>  H0: mean = 0
#>  alpha		= 0.01 
#>  p-value	= 2.32e-26
#>  Est. mean	= -1.229
#>  CI{1-alpha}	= [ -1.254 -1.205 ]
#>  d		= -35.04722
#> 
#>  Test 3 : algo2 x algo4
#>  H0: mean = 0
#>  alpha		= 0.0125 
#>  p-value	= 5.706e-25
#>  Est. mean	= -0.2946
#>  CI{1-alpha}	= [ -0.3016 -0.2876 ]
#>  d		= -28.6847
#> 
#>  Test 4 : algo3 x algo4
#>  H0: mean = 0
#>  alpha		= 0.01667 
#>  p-value	= 1.808e-24
#>  Est. mean	= 0.9348
#>  CI{1-alpha}	= [ 0.9121 0.9575 ]
#>  d		= 26.68792
#> 
#>  Test 5 : algo1 x algo2
#>  H0: mean = 0
#>  alpha		= 0.025 
#>  p-value	= 3.647e-21
#>  Est. mean	= 0.2864
#>  CI{1-alpha}	= [ 0.276 0.2968 ]
#>  d		= 16.57025
#> 
#>  ----- Stop rejecting H0 at this point -----
#> 
#>  Test 6 : algo1 x algo4
#>  H0: mean = 0
#>  alpha		= 0.05 
#>  p-value	= 0.05503
#>  Est. mean	= -0.008165
#>  CI{1-alpha}	= [ -0.01653 0.0001982 ]
#>  d		= -0.5019696
#> #====================================

# You can override some defaults if you want:
summary(my.results, test = "wilcoxon")

#> #====================================
#>  CAISEr object:
#>  Number of instances sampled: 17
#>  Number of instances required: 17
#>  Adequate power: TRUE
#>  Total runs of algo1 : 340
#>  Total runs of algo2 : 340
#>  Total runs of algo3 : 542
#>  Total runs of algo4 : 745
#> #====================================
#>  Pairwise comparisons of interest:
#>  Test: wilcoxon
#>  H1: two.sided
#>  Comparisons: all.vs.all
#>  Alpha (FWER): 0.05
#>  Power target: mean
#>  Desired power: 0.9
#> #====================================
#> Tests using Holm's step-down procedure:
#> 
#>  Test 1 : algo1 x algo2
#>  H0: median = 0
#>  alpha		= 0.008333 
#>  p-value	= 1.526e-05
#>  Est. median	= 0.2906
#>  CI{1-alpha}	= [ 0.2891 0.2913 ]
#>  d		= 16.57025
#> 
#>  Test 2 : algo1 x algo3
#>  H0: median = 0
#>  alpha		= 0.01 
#>  p-value	= 1.526e-05
#>  Est. median	= -0.9497
#>  CI{1-alpha}	= [ -0.9527 -0.9363 ]
#>  d		= -50.02262
#> 
#>  Test 3 : algo2 x algo3
#>  H0: median = 0
#>  alpha		= 0.0125 
#>  p-value	= 1.526e-05
#>  Est. median	= -1.241
#>  CI{1-alpha}	= [ -1.244 -1.226 ]
#>  d		= -35.04722
#> 
#>  Test 4 : algo2 x algo4
#>  H0: median = 0
#>  alpha		= 0.01667 
#>  p-value	= 1.526e-05
#>  Est. median	= -0.2924
#>  CI{1-alpha}	= [ -0.3024 -0.2897 ]
#>  d		= -28.6847
#> 
#>  Test 5 : algo3 x algo4
#>  H0: median = 0
#>  alpha		= 0.025 
#>  p-value	= 1.526e-05
#>  Est. median	= 0.9476
#>  CI{1-alpha}	= [ 0.9243 0.9532 ]
#>  d		= 26.68792
#> 
#>  ----- Stop rejecting H0 at this point -----
#> 
#>  Test 6 : algo1 x algo4
#>  H0: median = 0
#>  alpha		= 0.05 
#>  p-value	= 0.1202
#>  Est. median	= -0.002008
#>  CI{1-alpha}	= [ -0.01374 0.0002181 ]
#>  d		= -0.5019696
#> #====================================